% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addDimNames}
\alias{addDimNames}
\title{Add dimnames to a matrix}
\usage{
addDimNames(mat, letter)
}
\arguments{
\item{mat}{The matrix}

\item{letter}{A letter to start the rownames}
}
\value{
A matrix with rownames
}
\description{
Add dimnames to a matrix
}
