% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addFeatureColumn}
\alias{addFeatureColumn}
\title{Add columns with feature names, and remove rownames}
\usage{
addFeatureColumn(x)
}
\arguments{
\item{x}{A results matrix}
}
\value{
The matrix extended with two columns of feature names in front
}
\description{
Add columns with feature names, and remove rownames
}
