% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpScoreTest.R
\name{arrayDeriv}
\alias{arrayDeriv}
\title{Construct the nxnxg/2 array of derivatives for a nxn matrix to the g/2 covariance matrix parameters.}
\usage{
arrayDeriv(fittedGP, distMat, what)
}
\arguments{
\item{fittedGP}{The fitted Gaussian process (vector of 4 parameters)}

\item{distMat}{The distance matrix of Cx and Ey}

\item{what}{For which parameter is the derivative required?}
}
\value{
The matrix of derivatives
}
\description{
Construct the nxnxg/2 array of derivatives for a nxn matrix to the g/2 covariance matrix parameters.
}
