% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayMatProd}
\alias{arrayMatProd}
\title{Multiply an array with a matrix}
\usage{
arrayMatProd(A, M)
}
\arguments{
\item{A}{An mxmxp array}

\item{M}{An mxm matrix}
}
\value{
An mxmxp array
}
\description{

}
\details{
The speedup comes from a single call to %*%, very efficient in BLAS.
}
