% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayProd}
\alias{arrayProd}
\title{Multiply two arrays of same dimensions along the first two dimensions}
\usage{
arrayProd(A, B)
}
\arguments{
\item{A, B}{mxmxp arrays}
}
\description{
Array resulting from all matrix products of slices of array A mxmxp
with those of array B of the same dimensions,
}
