% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayProd2tr}
\alias{arrayProd2tr}
\title{Find traces of all inner products of matrices composing arrays A and B yielding a pxp matrix}
\usage{
arrayProd2tr(A, B)
}
\arguments{
\item{A, B}{mxmxp arrays}
}
\value{
pxp matrix of traces
}
\description{
A faster version of \code{vapply(seq_len(p), FUN.VALUE = double(p), function(i)\{
vapply(seq_len(p), FUN.VALUE = double(1), function(j)\{
   tr(arr[,,i] \%*\% arr2[,,j])
 \})
\})}
}
