% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayProdTr}
\alias{arrayProdTr}
\title{Find traces of all inner products of matrices composing arrays A and B, after transposing the second yielding a pxp matrix}
\usage{
arrayProdTr(A, B)
}
\arguments{
\item{A, B}{mxmxp arrays}
}
\value{
A pxp matrix of traces
}
\description{
Returns the matrix resulting from the traces of all matrix products of slices of array A mxmxp
with those of array B of the same dimensions. It is a faster version of
vapply(seq_len(p), FUN.VALUE = double(p), function(i){
     vapply(seq_len(p), FUN.VALUE = double(1), function(j){
         tr(tcrossprod(arr[,,i], arr2[,,j]))
     })
})
}
\details{
The speedup comes from a single call to crossprod, very efficient in BLAS.
}
