% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assayT}
\alias{assayT}
\title{Extract an assay, and transpose}
\usage{
assayT(x, assayName)
}
\arguments{
\item{x}{The SummarizedExperiment object}

\item{assayName}{The name of the assay}
}
\value{
The required assay, transposed
}
\description{
Extract an assay, and transpose
}
