% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bdiagn}
\alias{bdiagn}
\title{A wrapper for Matrix::bdiag maintaining names}
\usage{
bdiagn(A, B)
}
\arguments{
\item{A, B}{Matrix to be used in \link[Matrix]{bdiag}}
}
\value{
Same as \link[Matrix]{bdiag} but with dimnames
}
\description{
A wrapper for Matrix::bdiag maintaining names
}
