% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGP.R
\name{buildDerivArray}
\alias{buildDerivArray}
\title{Build a 3-slice array of covariance-parameter derivatives}
\usage{
buildDerivArray(fittedGP, distMat, suffix)
}
\arguments{
\item{fittedGP}{The fitted Gaussian process (vector of 4 parameters)}

\item{distMat}{Distance matrix for the relevant modality}

\item{suffix}{Character suffix to append to parameter names ("X" or "Y")}
}
\value{
An \eqn{n \times n \times 3} array with 3rd-dim names
  \code{c("sigma<suffix>", "range<suffix>", "nugget<suffix>")}
}
\description{
Build a 3-slice array of covariance-parameter derivatives
}
