% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCT.R
\name{CCT}
\alias{CCT}
\title{An analytical p-value combination method using the Cauchy distribution.}
\usage{
CCT(pvals)
}
\arguments{
\item{pvals}{a numeric vector of p-values, where each of the elements is
between 0 to 1, to be combined.}
}
\value{
The aggregated p-value
}
\description{
The \code{CCT} function takes in a numeric vector of p-values,
and returns the aggregated p-value using Cauchy combination rule.
The code was taken from the xihaoli/STAAR github repo \insertCite{Liu2020}{sbivar}, and adapted.
}
\references{
\insertAllCited{}
}
