% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransISingle.R
\name{evalVariogram}
\alias{evalVariogram}
\title{Evaluate a variogram on a set of distances}
\usage{
evalVariogram(vg, distVec)
}
\arguments{
\item{vg}{The variogram model (a \link[gstat]{variogramModel})}

\item{distVec}{A vector of pairwise distances}
}
\value{
A vector of covariances
}
\description{
Evaluate a variogram on a set of distances
}
