% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreWeights.R
\name{exploreWeights}
\alias{exploreWeights}
\title{Visualize different weighting functions}
\usage{
exploreWeights(
  etas,
  dists = seq(0, 0.1, length.out = 1000),
  palette = "paired"
)
}
\arguments{
\item{etas}{A vector of positive decay parameters}

\item{dists}{A set of distances, smaller than the square root of 2 since all coordinates are scaled to the unit square}

\item{palette}{The colour palette}
}
\value{
Plots the weighting functions
}
\description{
The weighting functions resulting from different choices of decay parameters eta are visualized in a lineplot
}
\examples{
exploreWeights(10^c(-5, -4, -3))
}
