% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{findDoubleUnderScore}
\alias{findDoubleUnderScore}
\title{Is there any double underscore in the character vector?}
\usage{
findDoubleUnderScore(charVec)
}
\arguments{
\item{charVec}{The character vector}
}
\value{
A boolean
}
\description{
Is there any double underscore in the character vector?
}
