% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGAM.R
\name{fitGAM}
\alias{fitGAM}
\title{Fit a GAM model to a single variable}
\usage{
fitGAM(df, outcome, family = gaussian(), Gamm, correlation)
}
\arguments{
\item{df}{The dataframe containing outcome and coordinates}

\item{outcome}{A character vector indicating the outcome variable}

\item{family}{A character string indicating the family, see \link[stats]{family}.}

\item{Gamm}{A boolean, should a gamm with Gaussian covariance kernel be fitted? See \link[mgcv]{gamm}}

\item{correlation}{Correlation structure, passed onto \link{fitGP} or \link{fitGAM}}
}
\value{
A fitted GAM model, or try-error when the fit fails
}
\description{
Fit a generalized additive model (GAM) captures spatial dependence through a bivariate spatial spline.
}
\details{
If a gamma fit is attempted and fails, which frequently happens for sparse data,
a negative binomial fit is attempted instead
}
\seealso{
\link[mgcv]{gam}, \link[mgcv]{s}
}
