% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGPs.R
\name{fitGP}
\alias{fitGP}
\title{Fit a Gaussian process (GP) to a single outcome vector}
\usage{
fitGP(x, coord, GPmethod, correlation, optControl)
}
\arguments{
\item{x}{outcome vector}

\item{coord}{Coordinate matrix}

\item{GPmethod}{The method by which to fit the Gaussian processes,
passed onto \link[nlme]{gls} as "method".}

\item{correlation}{The correlation object, see \link[nlme]{correlation}.
At this point, only \link[nlme]{corGaus} is accepted.}

\item{optControl}{List of control values, see \link[nlme]{glsControl}.}
}
\value{
A vector of length 4 with components range, nugget, sigma and mean
}
\description{
A Gaussian process (GP) models the spatial autocovariance using a kernel function
that describes the covariance as a decreasing function of distance between observations.
}
\note{
Fitting GPs can be computation and memory intensive!
}
