% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGPs.R
\name{fitGPs}
\alias{fitGPs}
\title{Title}
\usage{
fitGPs(x, Cx, GPmethod, training_iter, device, corStruct, optControl)
}
\arguments{
\item{x}{outcome vector}

\item{GPmethod}{The method by which to fit the Gaussian processes}

\item{training_iter}{Number of training iterations in gpytorch}

\item{device}{Device on which the GPs are fitted using the python package \textit{gpytorch}}

\item{corStruct}{The correlation object, see \link[nlme]{corStruct}}

\item{optControl}{List of control values, see \link[nlme]{glsControl}}
}
\value{
A vector of length 4 with components range, nugget, sigma and mean
}
\description{
Title
}
\details{
Providing "cuda" as device exploits the available GPU for speedup,
especially for large GPs (thousands of observations). Setting GPmethod to
 "ML" or "REML" leads to model fitting in R using the \link[nlme]{gls} function.
Setting GPmethod to "gpytorch" uses the eponymous python package, with the option of gpu accceleration.
The parametrization of the GPs is the one common in geostatistics, as described in \link[nlme]{gls}.
}
