% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLinModels.R
\name{fitLinModel}
\alias{fitLinModel}
\title{Fit a linear model for an individual feature pair}
\usage{
fitLinModel(Formula, dff, Control, MM, Weight = NULL)
}
\arguments{
\item{Formula}{A formula for the linear model to be fitted, can contain random effects}

\item{dff}{The dataframe}

\item{Control}{Control parameters}

\item{MM}{A boolean, should a mixed model be tried}

\item{Weight}{A weight variable}
}
\value{
A fitted model
}
\description{
Fit a linear model for an individual feature pair
}
\details{
Code is based on smoppix:::fitPiModel, but may diverge
}
\seealso{
\link{fitLinModels}
}
