% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGAM.R
\name{fitPPP}
\alias{fitPPP}
\title{Fit a poisson point process with splines}
\usage{
fitPPP(PPP)
}
\arguments{
\item{PPP}{a ppp object}
}
\value{
a ppm object
}
\description{
Fit a poisson point process with splines
}
