% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMsMulti.R
\name{GAMsMulti}
\alias{GAMsMulti}
\title{Fit GAMs and find correlations and standard error for data lists}
\usage{
GAMsMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families,
  n_points_grid,
  verbose,
  normX,
  normY,
  featuresX,
  featuresY,
  findVariances = FALSE
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families, n_points_grid}{See \link{GAMsSingle}}

\item{verbose}{Should info on type of analysis be printed?}

\item{featuresX, featuresY}{Features to be tested. Defaults to all features, but specifying them allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization.}

\item{findVariances}{Should variances be calculated? For internal use only}
}
\value{
A list named like Xl, containing all results
}
\description{
Wraps \link{GAMsSingle} for lists
}
