% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMsSingle.R, R/GAMsSinglePPP.R
\name{GAMsSingle}
\alias{GAMsSingle}
\title{Fit univariate GAMs and test bivariate combinations}
\usage{
GAMsSingle(
  X,
  Y,
  Cx,
  Ey,
  families,
  n_points_grid,
  verbose,
  featuresX,
  featuresY,
  Gamm,
  correlation,
  findVariances = TRUE
)

GAMsSingle(
  X,
  Y,
  Cx,
  Ey,
  families,
  n_points_grid,
  verbose,
  featuresX,
  featuresY,
  Gamm,
  correlation,
  findVariances = TRUE
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{families}{A vector of length 2 giving the distributional families
for the outcome values. See details of \link{sbivarSingle}.}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{verbose}{Should info on type of analysis be printed?}

\item{featuresX, featuresY}{Features to be tested. Defaults to all features, but specifying them allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization.}

\item{Gamm}{A boolean, should a gamm with Gaussian covariance kernel be fitted? See \link[mgcv]{gamm}}

\item{correlation}{Correlation structure, passed onto \link{fitGP} or \link{fitGAM}}

\item{findVariances}{Should variances be calculated? For internal use only}
}
\value{
A named list of results

A named list of results
}
\description{
Fit univariate GAMs with 2D cubic smoothing splines for both modalities, and test for correlation between all bivariate combinations.

Fit univariate GAMs with 2D cubic smoothing splines for both modalities, and test for correlation between all bivariate combinations.
}
