% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSigmas.R
\name{GaussKernel}
\alias{GaussKernel}
\title{Construct the SAC part of the covariance matrix using the Gaussian covariance kernel,}
\usage{
GaussKernel(distMat, range)
}
\arguments{
\item{distMat}{The distance matrix of Cx and Ey}

\item{range}{Range (or length-scale) parameter of the Gaussian covariance kernel}
}
\value{
The SAC covariance matrix
}
\description{
Construct the SAC part of the covariance matrix using the Gaussian covariance kernel,
}
