% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSigmas.R
\name{GaussKernelNugget}
\alias{GaussKernelNugget}
\title{Construct a covariance matrix using the Gaussian covariance kernel,
plus the variances on the diagonal (the nugget).}
\usage{
GaussKernelNugget(distMat, range, nugget)
}
\arguments{
\item{distMat}{The distance matrix of Cx and Ey}

\item{range, nugget}{Range and nugget parameters of the Gaussian covariance kernel}
}
\value{
The full covariance matrix
}
\description{
Construct a covariance matrix using the Gaussian covariance kernel,
plus the variances on the diagonal (the nugget).
}
\seealso{
\link[nlme]{corGaus}, \link[nlme]{corMatrix}
}
