% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAvDists}
\alias{getAvDists}
\title{Find average distances of bins}
\usage{
getAvDists(distMat, cutoff, width)
}
\arguments{
\item{distMat}{The distance matrix}

\item{width}{Width of the variogram estimation, passed onto \link[gstat]{variogram}}
}
\value{
A vector of indices
}
\description{
Find average distances of bins
}
\details{
tapply(distMat, bin_index, mean) Matches with the results from
\link[gstat]{variogram}, but this result is not needed here.
}
