% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getDiscreteVars}
\alias{getDiscreteVars}
\title{Get all categrocial variables from a dataframe}
\usage{
getDiscreteVars(df)
}
\arguments{
\item{df}{The data frane}
}
\value{
A character vector of variable names
}
\description{
Get all categrocial variables from a dataframe
}
