% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFeaturesList}
\alias{getFeaturesList}
\title{Return feature names from a list}
\usage{
getFeaturesList(Xl)
}
\arguments{
\item{Xl}{A list of feature matrices}
}
\value{
A vector of feature names
}
\description{
Return feature names from a list
}
