% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFeaturesX}
\alias{getFeaturesX}
\title{Retreive unique features names from a matrix or ppp}
\usage{
getFeaturesX(X)
}
\arguments{
\item{X}{the matrix or ppp object}
}
\value{
A vector of unique features
}
\description{
Retreive unique features names from a matrix or ppp
}
