% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getNormFun}
\alias{getNormFun}
\title{Return a normalization function based on a character string}
\usage{
getNormFun(norm)
}
\arguments{
\item{norm}{The character string}
}
\value{
a normalization function
}
\description{
Return a normalization function based on a character string
}
