% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSpatialCoords}
\alias{getSpatialCoords}
\title{Extract coordinate matrix}
\usage{
getSpatialCoords(X, Cx)
}
\arguments{
\item{X}{The matrix or SpatialExperiment object}

\item{Cx}{The coordinate matrix}
}
\value{
A coordinate matrix
}
\description{
Extract coordinate matrix
}
