% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getX}
\alias{getX}
\title{Extract data matrix}
\usage{
getX(X, assay)
}
\arguments{
\item{X}{The matrix or SpatialExperiment object}

\item{assay}{The name of the assay}
}
\value{
A matrix
}
\description{
Extract data matrix
}
