% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{giveValidNames}
\alias{giveValidNames}
\title{Give a matrix valid names}
\usage{
giveValidNames(X)
}
\arguments{
\item{X}{the matrix}
}
\value{
The same matrix, with potentially adapted column names
}
\description{
Give a matrix valid names
}
