% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpScoreTest.R
\name{gpScoreTest}
\alias{gpScoreTest}
\title{Perform the score test for significance of the bivariate spatial association for Gaussian processes.}
\usage{
gpScoreTest(
  x,
  y,
  Cx,
  Ey,
  device = "cpu",
  altSigmas,
  method = "gaussian",
  distMat = as.matrix(stats::dist(rbind(Cx, Ey))),
  solXonly,
  solYonly,
  sx,
  numLscAlts = if (missing(altSigmas)) 10 else dim(altSigmas)[3],
  Quants = c(0.005, 0.5)
)
}
\arguments{
\item{x, y}{outcome vectors}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{device}{Device to fit the GPs on. Defaults to "cpu",
can use "gpu" via the python package gpytorch if installed}

\item{altSigmas}{A prepared series of bivariate association matrices}

\item{method}{passed onto \link[nlme]{gls}}

\item{distMat}{The distance matrix of Cx and Ey}

\item{solXonly, solYonly}{Parametersof the Gaussian process}

\item{sx}{Inverse of the covariance matrix of x. Will be calculated if missing.}

\item{numLscAlts}{Number of length scales (and thus number of covariance matrices to be tested)}

\item{Quants}{Most extreme quantiles of the distance distribution to be used as length scales.}
}
\value{
A vector of length 2: a p-value and an indicator of the sign:
+1 for positive association, -1 for negative
}
\description{
Perform the score test for significance of the bivariate spatial association for Gaussian processes.
}
