% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPsSingle.R
\name{GPsSingle}
\alias{GPsSingle}
\title{Fit Gaussian processes (GPs) if needed, and perform score tests}
\usage{
GPsSingle(
  X,
  Y,
  Cx,
  Ey,
  gpParams,
  numLscAlts,
  Quants,
  GPmethod,
  corStruct,
  optControl,
  verbose,
  featuresX,
  featuresY
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{gpParams}{Parameters of the Gaussian processes, see details}

\item{numLscAlts}{Number of length scales to be tested for bivariate association}

\item{Quants}{Most extreme quantiles of the distance distribution to take as length scales}

\item{GPmethod}{The method by which to fit the Gaussian processes,
passed onto \link[nlme]{gls} as "method".}

\item{corStruct}{The correlation object, see \link[nlme]{corStruct}.
At this point, only \link[nlme]{corGaus} is accepted.}

\item{optControl}{List of control values, see \link[nlme]{glsControl}.}

\item{verbose}{Should info on type of analysis be printed?}

\item{featuresX, featuresY}{Features to be tested. Defaults to all features, but specifying them allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization.}
}
\value{
A named list of results
}
\description{
Fit all univariate GPs on both modalities, and perform all bivariate tests across them
}
\details{
gpParams must be a list of length 2 with names 'X' and 'Y', consisting of matrices
with rownames "mean", "nugget", "range" and "sigma", and column names as in X and Y.
This argument allows to pass parameters of the Gaussian processes estimated with other software
(e.g. with GPU acceleration) to perform the score test.
}
