% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeNames}
\alias{makeNames}
\title{Make unique names}
\usage{
makeNames(featX, featY)
}
\arguments{
\item{featX, featY}{vectors of feature names}
}
\value{
A vector of names
}
\description{
Make unique names
}
