% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{makeOffset}
\alias{makeOffset}
\title{Make a list of offsets}
\usage{
makeOffset(X, family)
}
\arguments{
\item{X}{The data matrix}

\item{family}{The distribution family}
}
\value{
A list of length two with offsets
}
\description{
Make a list of offsets
}
