% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapMoransI.R
\name{makeSigmaMatheron}
\alias{makeSigmaMatheron}
\title{Build a spatial autocovariance matrix based on a variogram object}
\usage{
makeSigmaMatheron(variogram, distId, n)
}
\arguments{
\item{variogram}{The variogram, estimated by \link[gstat]{variogram}}

\item{distance}{The distance matrix}
}
\value{
The autocovariance matrix
}
\description{
Build a spatial autocovariance matrix based on a variogram object
}
