% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransIMulti.R
\name{MoransIMulti}
\alias{MoransIMulti}
\title{Find all Moran's I statistics for a list of matrices}
\usage{
MoransIMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  findVariances,
  verbose,
  featuresX,
  featuresY,
  findMaxW,
  normX,
  normY,
  ...
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{findVariances}{Should variances be calculated? For internal use only}

\item{verbose}{Should info on type of analysis be printed?}

\item{featuresX, featuresY}{Features to be tested. Defaults to all features, but specifying them allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization.}

\item{findMaxW}{Is the maximum bivariate Moran's I needed?}

\item{...}{passed onto \link{MoransISingle}}
}
\value{
A list of Moran's I estimates, standard errors and maximum values
}
\description{
Find all Moran's I values for lists of observations matrices from different modalities,
by calling the \link{MoransISingle} function.
}
\seealso{
\link{MoransISingle}
}
