% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{moveCoords}
\alias{moveCoords}
\title{Shift coordinates to baseline 0}
\usage{
moveCoords(Coord)
}
\arguments{
\item{Coord}{coordinate matrix}
}
\value{
Shifted coordinate matrix
}
\description{
Shift coordinates to baseline 0
}
