% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{moveTwoCoords}
\alias{moveTwoCoords}
\title{Move two sets of coordinates to 0-1. without shifting them with respect to each other}
\usage{
moveTwoCoords(Cx, Ey)
}
\arguments{
\item{Cx, Ey}{Coordinate matrices}
}
\value{
A list with the shifted and scaled coordinates
}
\description{
Move two sets of coordinates to 0-1. without shifting them with respect to each other
}
