% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normMat}
\alias{normMat}
\title{Normalize a data matrix}
\usage{
normMat(x, norm, pseudoCount = 1e-08)
}
\arguments{
\item{x}{The matrix}

\item{norm}{A character string, either "none", "log" or "rel"}

\item{pseudoCount}{A pseudocount added prior to log-normalization to avoid taking the log of zero}
}
\value{
A normalized matrix
}
\description{
Normalize to relative expression, and potentially add pseudocount and log-normalize.
}
\details{
norm = "none" is pass-through, norm = "rel" divides by sample sums,
"log" adds a pseudocount, divides by sample sums and log-normalizes.
}
\examples{
mat = matrix(rpois(2000, lambda = 3), 40, 50)
nMat = normMat(mat, norm = "rel")
}
