% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoords.R
\name{plotCoords}
\alias{plotCoords}
\title{Plot the coordinates of two omics modalities}
\usage{
plotCoords(Cx, Ey, pch = 20, cex = 6, pchY = pch, ...)
}
\arguments{
\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{pch, pchY}{Point shapes for x and y}

\item{cex}{Expansion factor}

\item{...}{passed onto plot()}
}
\value{
Plots to the plotting device
}
\description{
Plot the coordinates of two modalities onto the same coordinate framework, in two different colours.
This is a useful check of the alignment and overlap
}
\examples{
example(sbivarSingle, "sbivar")
plotCoords(Cx, Ey)
For multiple coordinates
example(sbivarMulti, "sbivar")
par(mfrow = c(2,3))
foo = lapply(names(Cxl), function(nam){
plotCoords(Cxl[[nam]], Eyl[[nam]], main = nam)
})
par(mfrow=c(1,1))
}
