% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairPPP.R
\name{plotPairSinglePPvec}
\alias{plotPairSinglePPvec}
\title{Plot a feature pair consisting of a single molecule point pattern and a quantitative outcome}
\usage{
plotPairSinglePPvec(
  Cx,
  y,
  Ey,
  sizeX = 0.1,
  sizeY = 1,
  shapeX = 15,
  shapeY = 1,
  modalityNames = c("Modality X", "Modality Y"),
  theme = theme_bw(),
  ...
)
}
\arguments{
\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{modalityNames}{Names to be given to the modalities,
appearing in the strip text of the columns. For plotTopPair() and
plotPairSingle(), the feature names are used.}

\item{theme}{the ggplot2 theme}

\item{...}{passed onto lower level functions}
}
\value{
A ggplot object
}
\description{
Plot a chosen feature pair, or the highest ranking feature pair,
for a single image
}
\examples{
n <- 8e1
m <- 9e1
p <- 3
k <- 2
Y <- matrix(rnorm(m * k), m, k,
    dimnames =
        list(paste0("sampleY", seq_len(m)), paste0("Y", seq_len(k)))
)
Ey <- matrix(runif(m * 2), m, 2, dimnames = list(rownames(Y), c("x", "y")))
# Single image analysis on synthetic point pattern + quantitative modality
library(spatstat.random)
lambda <- 8e1
p <- 3
PPP <- rmpoispp(lambda, types = paste0("gene", seq_len(p)), win = owin(c(0, 1), c(0, 1)))
marks(PPP, drop = FALSE) <- data.frame(feature = marks(PPP)) # Make sure marks is a dataframe
resMoransIppp <- sbivar(X = PPP, Y, Ey, method = "Moran's I")
plotTopPairSinglePPP(resMoransIppp, PPP = X, Y = Y, Ey = Ey)
}
\seealso{
\link{plotTopPair}, \link{sbivarSinglePPP}
}
