% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPair.R
\name{plotPairSingleVectors}
\alias{plotPairSingleVectors}
\title{Plot a feature pair}
\usage{
plotPairSingleVectors(
  x,
  y,
  Cx,
  Ey,
  modalityNames = c("feature 1", "feature 2"),
  size = 2
)
}
\arguments{
\item{x, y}{Outcome vectors}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{resultsSingle, resultsMulti}{Results, returned by \link{sbivarSingle} and
\link{extractResultsMulti}, respectively.}

\item{featx, featy}{The features in X and Y to be plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.}
}
\value{
A ggplot2 object
}
\description{
Plot a chosen feature pair for a single or list of matrices,
or plot the highest ranking feature pair.
}
\examples{
example(sbivarSingle, "sbivar")
plotPairSingle(X[,1], Y[, 1], Cx, Ey, features = c("X1", "Y1"))
plotTopResultsSingle(resModtTest, Cx, Ey)
plotPairMulti(Xl, Yl, Cxl, Eyl, featx = "X1", featy = "Y1")
plotTopResultsMulti(estCorrelations, Xl, Yl, Cxl, Eyl)
}
