% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpotSums.R
\name{plotSpotSums}
\alias{plotSpotSums}
\title{Plot the spot-wise sums of the two modalities across the slices}
\usage{
plotSpotSums(
  Xl,
  Yl,
  Cxl,
  Eyl,
  viewNames = c("Transcriptomics", "Metabolomics"),
  pointSize = 0.01,
  stripTextSize = 8.5
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{viewNames}{The names of the views, showed in the column names}

\item{pointSize}{The point size for geom_point()}

\item{stripTextSize}{The size of the text in the strips (grey panels)}
}
\value{
A ggplot2 object
}
\description{
This is an important exploratory plot to visualize spot-wise sums
(library sizes, total ion counts etc.) spatially
}
\examples{
data(Vicari)
plotSpotSums(
    Vicari$TranscriptOutcomes, Vicari$MetaboliteOutcomes,
    Vicari$TranscriptCoords, Vicari$MetaboliteCoords
)
}
