% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairMulti.R
\name{plotTopResultsMulti}
\alias{plotTopResultsMulti}
\title{Plot a chosen or top feature in the analysis for a series of images}
\usage{
plotTopResultsMulti(resultsMulti, Xl, Yl, Cx, Ey, parameter = "Intercept")
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.}
}
\description{
Plot a chosen or top feature in the analysis for a series of images
}
\examples{
example(fitLinModels, "sbivar")
plotTopResultsMulti(resGams, Xl, Yl, Cxl, Eyl)
plotPairMulti(Xl, Yl, Cxl, Eyl, feauresx = c("X1", "Y1"))
}
