% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printIteration}
\alias{printIteration}
\title{Print a message for the current iteration}
\usage{
printIteration(current, all)
}
\arguments{
\item{current}{current state of the iterator}

\item{all}{vector of all iterators}
}
\value{
prints message to output
}
\description{
Print a message for the current iteration
}
