% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printProgress}
\alias{printProgress}
\title{Print feature progress}
\usage{
printProgress(feat, allFeats, verbose)
}
\arguments{
\item{feat}{The current feature}

\item{allFeats}{Vector of all features}

\item{verbose}{Boolean, should output be printed?}
}
\value{
Prints progress message to output
}
\description{
Print feature progress
}
