% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replaceLhs}
\alias{replaceLhs}
\title{Replace the left hand side of a formula by a fixed string}
\usage{
replaceLhs(x, repl = "out")
}
\arguments{
\item{x}{a formula}

\item{repl}{the replacement string}
}
\value{
A formula
}
\description{
Replace the left hand side of a formula by a fixed string
}
