% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-sbivar.R
\name{sbivar}
\alias{sbivar}
\alias{sbivar,matrix-method}
\alias{sbivar,list-method}
\alias{sbivar,SpatialExperiment-method}
\alias{sbivar,MultiAssayExperiment-method}
\title{Spatial bivariate association analysis}
\usage{
sbivar(X, ...)

\S4method{sbivar}{matrix}(X, Y, Cx, Ey, ...)

\S4method{sbivar}{list}(X, Y, Cx, Ey, assayX = NULL, assayY = NULL, ...)

\S4method{sbivar}{SpatialExperiment}(X, Y, assayX, assayY, sample_id_x, sample_id_y = sample_id_x, ...)

\S4method{sbivar}{MultiAssayExperiment}(X, experimentX, experimentY, assayX, assayY, ...)
}
\arguments{
\item{X}{the input object, containing measurements of the first modality, see methods('sbivar')}

\item{...}{additional constructor and analysis arguments}

\item{Y}{Matrix or SpatialExperiment object of second modality}

\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{assayX, assayY}{Assay names to be used in the analysis,
see \link[SummarizedExperiment]{assay}}

\item{sample_id_x, sample_id_y}{If provided, these are used to discriminate between different images
included in the same SpatialExperiment objects X and Y. By default, they are assumed to be the same for both X and Y.}

\item{experimentX, experimentY}{Names of the experiments in X and Y to be used in the analysis}
}
\value{
A list containing the analysis result, along with parameters used in the analysis
}
\description{
Perform a bivariate spatial association analysis, either on a single or multiple images.
Depending on the input, the workhorse functions \link{sbivarSingle}
(single-image) or \link{sbivarMulti} (multi-image) are called.
}
\examples{
# Single image analysis on synthetic data
n <- 8e1
m <- 9e1
p <- 3
k <- 2
X <- matrix(rnorm(n * p), n, p,
    dimnames =
        list(paste0("sampleX", seq_len(n)), paste0("X", seq_len(p)))
)
Y <- matrix(rnorm(m * k), m, k,
    dimnames =
        list(paste0("sampleY", seq_len(m)), paste0("Y", seq_len(k)))
)
Cx <- matrix(runif(n * 2), n, 2, dimnames = list(rownames(X), c("x", "y")))
Ey <- matrix(runif(m * 2), m, 2, dimnames = list(rownames(Y), c("x", "y")))
resMoransI <- sbivar(X, Y, Cx, Ey, method = "Moran's I")
resGAMs <- sbivar(X, Y, Cx, Ey, method = "GAMs")
Y2 <- matrix(rnorm(n * k), n, k,
    dimnames =
        list(paste0("sampleX", seq_len(n)), paste0("Y", seq_len(k)))
)
resModtTestJoint <- sbivar(X, Y2, Cx, method = "Modified t-test")
# Single image analysis on synthetic data, converted to SpatialExperiment
if (require(SpatialExperiment)) {
    seX <- SpatialExperiment(
        assays = list("transcripts" = t(X)),
        spatialCoords = Cx
    )
    seY <- SpatialExperiment(
        assays = list("metabolites" = t(Y)),
        spatialCoords = Ey
    )
    resModtGPs <- sbivar(seX, seY,
        assayX = "transcripts", assayY = "metabolites",
        method = "GPs"
    )
}
}
\seealso{
\link{sbivarSingle}, \link{sbivarMulti}
}
