% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scaleHelpFun}
\alias{scaleHelpFun}
\title{Wrapper to normalize, select feature and scale}
\usage{
scaleHelpFun(X, feat, normFun)
}
\arguments{
\item{X}{data matrix}

\item{feat}{the feature name}

\item{normFun}{normalizing function}
}
\value{
A vector of values
}
\description{
Wrapper to normalize, select feature and scale
}
\details{
Returns vector of NA if feature not found, leading to grey in the plots
}
