% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitSpatialExperiment}
\alias{splitSpatialExperiment}
\title{Split a SpatialExperiment object into images}
\usage{
splitSpatialExperiment(spe, sample_id)
}
\arguments{
\item{spe}{The SpatialExperiment object}

\item{sample_id}{A character vector}
}
\value{
A list of SpatialExperiment objects
}
\description{
Split Spatial Experiment into a list of SpatialExperiment objects,
based on a variable present in the colData slot
}
