% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sund}
\alias{sund}
\title{Split a string}
\usage{
sund(string, split = "__")
}
\arguments{
\item{string}{The string}

\item{split}{string to split by}
}
\value{
A character vector of length 2
}
\description{
Split a string
}
