% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGAM.R
\name{testGAM}
\alias{testGAM}
\title{Test for correlation between the predictions of two GAM models}
\usage{
testGAM(modelx, modely, predx, predy)
}
\arguments{
\item{modelx, modely}{Two fitted GAMs}

\item{predx, predy}{Predictions and covariance matrices of fitted GAMs in common grid}
}
\value{
A vector with correlation, its standard error and the p-value
}
\description{
The variance of the correlation between the spline surfaces is found by
propagating the uncertainties on the spline parameters through uncertainty on
the spline predictions to uncertainty on the correlation estimate.
}
