% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGAM.R
\name{testManyGAMs}
\alias{testManyGAMs}
\title{Test all bivariate combinations for fitted lists of GAMs}
\usage{
testManyGAMs(gamsx, gamsy, newGrid)
}
\arguments{
\item{gamsx, gamsy}{Lists of GAM models for two modalities}

\item{newGrid}{The new grid of points in which to evaluate the GAMs}
}
\value{
A named list of results
}
\description{
Test all bivariate combinations for fitted lists of GAMs
}
