% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGP.R
\name{testManyGPs}
\alias{testManyGPs}
\title{Test all bivariate combinations for fitted lists of GAMs}
\usage{
testManyGPs(gpsx, gpsy, X, Y, distMat, numLscAlts, Quants)
}
\arguments{
\item{gpsx, gpssy}{Lists of GAM models for two modalities}

\item{newGrid}{The new grid of points in which to evaluate the GAMs}
}
\value{
A named list of results
}
\description{
Test all bivariate combinations for fitted lists of GAMs
}
