% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tr}
\alias{tr}
\title{Find trace of a matrix, of traces of an array}
\usage{
tr(x, dim = c(1, 2))
}
\arguments{
\item{x}{Matrix or array}

\item{dim}{Dimensions defining matrices to find traces over}
}
\value{
A trace or vector of traces
}
\description{
A (mxm) matric has one trace (the sum of the diagonal elements), a (mxmxp) array has p traces
}
