% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vicari}
\alias{Vicari}
\title{Spatial transcriptomics and metabolomics data of mouse brain}
\format{
Four lists of data matrices:
\describe{
  \item{TranscriptCoords,MetaboliteCoords}{Coordinate lists}
  \item{TranscriptOutcomes,MetaboliteOutcomes}{Outcome matrices}
}

Four lists of data matrices:
\describe{
  \item{TranscriptCoords,MetaboliteCoords}{Coordinate lists}
  \item{TranscriptOutcomes,MetaboliteOutcomes}{Outcome matrices}
}
}
\source{
\doi{10.1038/s41587-023-01937-y}

\doi{10.1038/s41587-023-01937-y}
}
\usage{
data(Vicari)

data(Vicari)
}
\description{
Spatial transcriptomics and metabolomics data measured on the same tissue sections
 of mouse brains on a regular grid by \insertCite{Vicari2024;nobrackets}{sbivar}. Only a subset of the data, consisting of
 the 5 most abundant transcripts and metabolites for 6 samples, are included in the package for computational and memory reasons.
 The images were pre-aligned manually with the help of MAGPIE \insertCite{Williams2025}{sbivar}.
The data consist of two lists of outcome variables and their coordinates.

Spatial transcriptomics and metabolomics data measured on the same tissue sections
 of mouse brains on a regular grid by \insertCite{Vicari2024;nobrackets}{sbivar}. Only a subset of the data, consisting of
 the 5 most abundant transcripts and metabolites for 6 samples, are included in the package for computational and memory reasons.
 The images were pre-aligned manually with the help of MAGPIE \insertCite{Williams2025}{sbivar}.
The data consist of two lists of outcome variables and their coordinates.
}
\references{
\insertAllCited{}

\insertAllCited{}
}
\keyword{datasets}
