% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapCorrelationsMulti.R
\name{wrapCorrelationsMulti}
\alias{wrapCorrelationsMulti}
\title{Find all cross-correlations for a list of matrices}
\usage{
wrapCorrelationsMulti(Xl, Yl, verbose)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{verbose}{Should info on type of analysis be printed?}
}
\value{
A list of named correlation vectors
}
\description{
Find all raw cross-correlations between lists of observations matrices from different modalities.
}
