% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapGAMsMulti.R
\name{wrapGAMsMulti}
\alias{wrapGAMsMulti}
\title{Wrapper function to fit GAMs and find correlations and standard error for data lists}
\usage{
wrapGAMsMulti(Xl, Yl, Cxl, Eyl, families, n_points_grid)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families}{A vector of length 2 giving outcome values.}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}
}
\value{
A list named like Xl, containing all results
}
\description{
Wrapper function to fit GAMs and find correlations and standard error for data lists
}
