% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapGPs.R
\name{wrapGPs}
\alias{wrapGPs}
\title{Fit Gaussian processes (GPs) if needed, and perform score tests}
\usage{
wrapGPs(
  X,
  Y,
  Cx,
  Ey,
  gpParams,
  numLscAlts,
  Quants,
  device,
  GPmethod,
  training_iter,
  corStruct,
  optControl
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{gpParams}{Parameters of the Gaussian processes, see details}

\item{numLscAlts}{Number of length scales to be tested for bivariate association}

\item{Quants}{Most extreme quantiles of the distance distribution to take as lenght scales}
}
\value{
A named list of results
}
\description{
Fit Gaussian processes (GPs) if needed, and perform score tests
}
