% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapModTtest.R
\name{wrapModTtest}
\alias{wrapModTtest}
\title{Perform modified t-tests for all pairs}
\usage{
wrapModTtest(X, Y, Cx, verbose)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx}{The shared coordinate matrix}

\item{verbose}{Should info on type of analysis be printed?}
}
\value{
A dataframe of results sorted by p-value, also containing effective sample size (ESS) and correlation estimate.
}
\description{
Then modified t-test is applied to all pairs, which tests for the significance
of the Pearson correlation while accounting for spatial autocorrelation \insertCite{Clifford1989}{sbivar}.
}
\references{
\insertAllCited{}
}
\seealso{
\link[SpatialPack]{modified.ttest}
}
