% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapMoransI.R
\name{wrapMoransI}
\alias{wrapMoransI}
\title{Calculate ibivariate Moran's I between two modality matrix, with variance and p-value}
\usage{
wrapMoransI(X, Y, Cx, Ey, wo, eta, numNN, cutoff, width, verbose, ...)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{wo, numNN, eta}{passed onto \link{buildWeightMat}}

\item{width}{Width of the variogram estimation, passed onto \link[gstat]{variogram}}

\item{verbose}{Should info on type of analysis be printed?}

\item{...}{passed onto \link[gstat]{variogram}}
}
\value{
A dataframe of results sorted by p-value, also containing the estimated Moran's I statistic and its variance.
}
\description{
The variance calculation requires estimation of the spatial autocorrelation structure of every feature separately
}
\references{
\insertAllCited{}
}
