% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapMoransIMulti.R
\name{wrapMoransIMulti}
\alias{wrapMoransIMulti}
\title{Find all Moran's I statistics for a list of matrices}
\usage{
wrapMoransIMulti(Xl, Yl, Cxl, Eyl, wo, numNN)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{wo}{The weighting option, either "nn" or "distance}

\item{numNN}{An integer, the number of neighbours}
}
\value{
A list of named Moran's I vectors
}
\description{
Find all raw Moran's I values between lists of observations matrices from different modalities.
Observation matrices are scaled and centered prior to calculation
}
\seealso{
buildWeightMat
}
