% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayMatProd}
\alias{arrayMatProd}
\title{Multiply an array with a matrix}
\usage{
arrayMatProd(A, M)
}
\description{
Array resulting from all matrix products of slices of array A mxmxp, with matrix M is mxm
}
