% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayMatProd}
\alias{arrayMatProd}
\title{Multiply an array with a matrix}
\usage{
arrayMatProd(A, M)
}
\arguments{
\item{A}{An mxmxp array}

\item{M}{An mxm matrix}
}
\value{
An mxmxp array
}
\description{
Array resulting from all matrix products of slices of array A mxmxp, with matrix M is mxm.
This is a faster version of \code{vapply(seq_len(p), FUN.VALUE = mat, function(i)\{mat \%*\% arr[,,i]\})},
although it may consume more memory
}
\details{
The speedup comes from a single call to %*%, very efficient in BLAS.
}
