% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayProd.R
\name{arrayProd2tr}
\alias{arrayProd2tr}
\title{Find traces of all inner products of matrices composing arrays A and B, yielding a pxp matrix}
\usage{
arrayProd2tr(A, B)
}
\arguments{
\item{A, B}{mxmxp arrays}
}
\value{
pxp matrix of traces
}
\description{
Find traces of all inner products of matrices composing arrays A and B, yielding a pxp matrix
}
