% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSigmas.R
\name{buildAltSigmas}
\alias{buildAltSigmas}
\title{Title}
\usage{
buildAltSigmas(distMat, numLscAlts, Quants, idN, idM)
}
\arguments{
\item{distMat}{The distance matrix of Cx and Ey}

\item{numLscAlts}{Number of length scales (and thus number of covariance matrices to be tested)}

\item{Quants}{Most extreme quantiles of the distance distribution to be used as length scales.}

\item{idN, idM}{indices for x and y in the distance matrix}
}
\value{
A list of covariance matrices
}
\description{
Title
}
