% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNewGrid.R
\name{buildNewGrid}
\alias{buildNewGrid}
\title{Build a new grid given two coordinate matrices}
\usage{
buildNewGrid(Cx, Ey, n_points_grid)
}
\arguments{
\item{Cx, Ey}{The coordinate matrices}

\item{n_points_grid}{an integer, the number of points desired for the new grid}
}
\value{
A data frame of two columns with all points of the grid,
with column names x and y.
}
\description{
Given two coordinate matrices, concave hulls are found around them. The intersection
between these two hulls is found, and in that area an evenly spaced, discrete
grid is constructed.
}
\details{
This function is mainly used to create a grid on which fitted GAMs can be
evaluated to calculate correlations.

The new grid will contain approximately the number of new points requested,
depending on the size of the concave hull
}
\examples{
Cx = matrix(runif(40, 0, 1), 20, 2)
Ey = matrix(runif(30, 0, 1), 15, 2)
buildNewGrid(Cx, Ey, n_points_grid = 50)
}
\seealso{
\link[concaveman]{concaveman}
}
