% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNewGrid.R
\name{buildNewGrid}
\alias{buildNewGrid}
\title{Build a new grid in the intersection of two coordinate matrices}
\usage{
buildNewGrid(Cx, Ey, n_points)
}
\arguments{
\item{Cx, Ey}{The coordinate matrices}

\item{n_points}{an integer, the number of new points}
}
\value{
A data frame of two columns with all points of the grid,
with columnnames x and y.
}
\description{
Given two coordinate matrices, concave hulls are found around them. The intersection
between these two hulls is found, and in that area an evenly spaced, discrete
grid is constructed.
}
\details{
This function is mainly used to create a grid on which fitted GAMs can be
evaluated to calculate correlations.
}
\examples{
Cx = matrix(runif(40, 0, 1), 20, 2)
Ey = matrix(runif(30, 0, 1), 15, 2)
buildNewGrid(Cx, Ey, n_points = 50)
}
