% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSigmas.R
\name{buildSigmaGp}
\alias{buildSigmaGp}
\title{Build the SAC matrix for a Gaussian process}
\usage{
buildSigmaGp(pars, distMat)
}
\arguments{
\item{pars}{A vector of parameters}

\item{distMat}{The distance matrix of Cx and Ey}
}
\value{
The covariance matrix
}
\description{
Build the SAC matrix for a Gaussian process
}
\seealso{
\link[nlme]{corGaus}, \link[nlme]{corMatrix}
}
