% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildWeightMat.R
\name{buildWeightMat}
\alias{buildWeightMat}
\title{Build a weight matrix for bivariate Moran's I}
\usage{
buildWeightMat(Cx, Ey, wo = c("distance", "nn"), numNN = 8)
}
\arguments{
\item{Cx, Ey}{Two coordinate matrices}

\item{wo}{The weighting option, either "nn" or "distance"}

\item{numNN}{An integer, the number of neighbours}
}
\value{
A weight matrix
}
\description{
Build a weight matrix to be used in the calculation of the bivariate Moran's I
statistic.
}
