% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkInputSingle}
\alias{checkInputSingle}
\title{Check input of matrices and lists}
\usage{
checkInputSingle(X, Y, Cx, Ey)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}
}
\value{
Throws error when a fault is found with the input, otherwise finishes silently
}
\description{
Checks dimensions of matrices and lists, to be used for all exported functions
}
