% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Duchini}
\alias{Duchini}
\title{Spatial transcriptomics and proteomics data of data of human hepatocellular carcinoma (HCC)}
\format{
A point pattern, and two matrices
\describe{
  \item{TranscriptsDuchini}{A point pattern of class ppp}
  \item{ProteinCoordsDuchini}{Protein coordinate matrix}
  \item{ProteinOutcomesDuchini}{Protein outcome matrix}
}
}
\source{
\doi{10.64898/2026.08.17.742355}
}
\usage{
data(Duchini)
}
\description{
Single-molecule spatial transcriptomics and lattice protein immunofluorescence data measured on the same tissue section
 of a human hepatocellular carcinoma sample by \insertCite{Duchini2026;nobrackets}{sbivar}. Only a subset of the data, consisting of
 two transcripts and two proteins is included in the package for computational and memory reasons.
 The images were pre-aligned by the authors
}
\references{
\insertAllCited{}
}
\keyword{datasets}
