% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransISingle.R
\name{evalVariogram}
\alias{evalVariogram}
\title{Evaluate a variogram on a distance matrix}
\usage{
evalVariogram(vg, distMat)
}
\arguments{
\item{vg}{The variogram}

\item{distMat}{The distance matrix}
}
\value{
A covariance matrix
}
\description{
Evaluate a variogram on a distance matrix
}
