% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGAM.R
\name{fitGAM}
\alias{fitGAM}
\title{Fit a GAM model to a single variable}
\usage{
fitGAM(
  df,
  outcome,
  family = gaussian(),
  offset = NULL,
  Gamm,
  correlation,
  gamMethod,
  bs
)
}
\arguments{
\item{df}{The dataframe containing outcome and coordinates}

\item{outcome}{A character vector indicating the outcome variable}

\item{family}{A character string indicating the family, see \link[stats]{family}.}

\item{offset}{A numeric vector with the offset, e.g. the library sizes for
spatial transcriptomic data, already on the scale of the regressor,
so log-transformed for count models.}

\item{correlation}{A corStruct object, see \link[nlme]{corStruct}.
At this point, \link[nlme]{corGaus} is hard-coded,}
}
\value{
A fitted GAM model, or try-error when the fit fails
}
\description{
Fit a generalized additive model (GAM) captures spatial dependence through a bivariate spatial spline.
}
\details{
If a gamma fit is attempted and fails, which frequenlty happens for sparse data,
a negative binomial fit is attempted instead
}
\seealso{
\link[mgcv]{gam},\link[mgcv]{s}
}
