% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGAM.R
\name{fitGAM}
\alias{fitGAM}
\title{Fit a GAM model to a single variable}
\usage{
fitGAM(df, outcome, k = -1, family = gaussian(), offset = NULL)
}
\arguments{
\item{df}{The dataframe containing outcome and}

\item{outcome}{A character vector indicating the outcome variable}

\item{k}{Dimension of the basis of the smooth term see \link[mgcv]{s}.}

\item{family}{A character string indicating the family, see \link[stats]{family}.}

\item{offset}{A numeric vector with the offset, e.g. the library sizes for
spatial transcriptomic data, already on the scale of the regressor,
so log-transformed for count models}
}
\value{
A fitted GAM model
}
\description{
Fit a GAM model to a single variable
}
