% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLinModels.R
\name{fitLinModel}
\alias{fitLinModel}
\title{Fit a linear model for an individual feature pair}
\usage{
fitLinModel(ff, y, Control, Terms, modMat, MM, Assign, weights = NULL)
}
\arguments{
\item{ff}{The prepared frame}

\item{y}{outcome vector}

\item{Control}{A control list for lmerTest::lmer}

\item{modMat}{Design matrix of the fixed effects model}

\item{MM}{a Boolean, should a mixed model fit be attempted?}

\item{Assign, Terms}{Added to fitted fixed effects model}

\item{weights}{weights vector}
}
\value{
A fitted model

A fitted lmer model
}
\description{
Fit a linear model for an individual feature pair
}
\details{
Code is based on smoppix:::fitSingleLmmModel, but may diverge
}
