% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResultsMulti.R, R/fitLinModels.R
\name{extractResults}
\alias{extractResults}
\alias{fitLinModels}
\title{Extract results from a list of fitted LMMs}
\usage{
extractResults(models, design, method = "BH")

fitLinModels(
  measures,
  design,
  Formula,
  Control = lmerControl(check.conv.grad = .makeCC("ignore", tol = 0.002, relTol = NULL),
    check.conv.singular = .makeCC(action = "ignore", tol = formals(isSingular)$tol),
    check.conv.hess = .makeCC(action = "ignore", tol = 1e-06))
)
}
\arguments{
\item{models}{The models}

\item{design}{A design dataframe}

\item{method}{Multiplicity correction method passed onto p.adjust}

\item{measures}{A list of measures of bivariate spatial association}

\item{Formula}{A formula for the linear model to be fitted, can contain random effects}

\item{Control}{A control list for lmerTest::lmer}
}
\value{
A list of matrices, all containing estimate, standard error,
p-value and adjusted p-value

A data frame containing, sorted by ascending p-value
}
\description{
The results are returned as matrix, including adjusted p-values, and sorted by .

Given measures estimated from replicated images using \link{sbivarMulti},
fit linear models to determine significance.
To maintain interpretability of the intercept, continuous fixed effect variables are centered,
and sum coding is used for the categorical ones.
}
\examples{
example(sbivarMulti, "sbivar")
toyDesign = data.frame("covariate" = rnorm(ims), "cofactor" = factor(rep(c(TRUE, FALSE), length.out = ims)),
"group" = rep(c("control", "treatment"), length.out = ims))
multiFitGams = fitLinModels(estGAMs, design = toyDesign, Formula = out ~ covariate + cofactor + (1|group))
multiFitMoran = fitLinModels(estMoran, design = toyDesign, Formula = out ~ covariate + cofactor)
multiFit = fitLinModels(estCorrelations, design = toyDesign, Formula = out ~ covariate + (1|group))
#Extract the results
resGams = extractResults(multiFitGams, design = toyDesign)
head(resGams$Intercept)
}
\seealso{
\link{fitLinModels}, \link[stats]{p.adjust}

\link[lmerTest]{lmer}, \link[stats]{lm}
}
