% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLinModels.R
\name{fitLinModels}
\alias{fitLinModels}
\title{Fit linear models on replicated image results}
\usage{
fitLinModels(
  measures,
  design,
  Formula,
  Control = lmerControl(check.conv.grad = .makeCC("ignore", tol = 0.002, relTol = NULL),
    check.conv.singular = .makeCC(action = "ignore", tol = formals(isSingular)$tol),
    check.conv.hess = .makeCC(action = "ignore", tol = 1e-06))
)
}
\arguments{
\item{measures}{A list of measures of bivariate spatial association}

\item{design}{A design dataframe}

\item{Formula}{A formula for the linear model to be fitted, can contain random effects}

\item{Control}{A control list for lmerTest::lmer}
}
\value{
A data frame containing, sorted by ascending p-value
}
\description{
Given measures estimated from replicated images, fit linear models to determine significance.
The measures are usually estimated using \link{sbivarMulti}
}
\examples{
example(sbivarMulti, "sbivar")
toyDesign = data.frame("covariate" = rnorm(ims), "group" = rep(c("control", "treatment"), length.out = ims))
multiFitGams = fitLinModels(estGAMs, design = toyDesign, Formula = out ~ covariate + (1|group))
multiFitMoran = fitLinModels(estMoran, design = toyDesign, Formula = out ~ covariate)
multiFit = fitLinModels(estCorrelations, design = toyDesign, Formula = out ~ covariate)
}
\seealso{
\link[lmerTest]{lmer}, \link[stats]{lm}
}
