% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLinModels.R
\name{fitLinModels}
\alias{fitLinModels}
\title{Fit linear models on replicated image results}
\usage{
fitLinModels(measures, design, Formula)
}
\arguments{
\item{measures}{A list of measures of bivariate spatial association}

\item{design}{A design dataframe}

\item{Formula}{A formula for the linear model to be fitted, can contain random effects}
}
\value{
A data frame containing, sorted by ascending p-value
}
\description{
Given measures estimated from replicated images, fit linear models to determine significance.
The measures are usually estimated using \link{sbivarMulti}
}
\examples{
example(sbivarMulti, "sbivar")
toyDesign = data.frame("covariate" = rnorm(im))
multiFitGams = fitLinModels(estGAMs, design = toyDesign,)
}
\seealso{
\link[lmerTest]{lmer}, \link[stats]{lm}
}
