% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGAM.R
\name{fitManyGAMs}
\alias{fitManyGAMs}
\title{Fit GAMs to all colums of a dataframe, as a wrapper for fitGAM}
\usage{
fitManyGAMs(mat, coord, family = gaussian(), ...)
}
\arguments{
\item{mat}{The matrix of outcomes}

\item{coord}{The coordinate matrix}

\item{family}{A character string indicating the family, see \link[stats]{family}.}

\item{...}{Passed onto \link{fitGAM}}
}
\value{
A list of GAM models
}
\description{
Fit GAMs to all colums of a dataframe, as a wrapper for fitGAM
}
