% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGPs.R
\name{fitManyGPs}
\alias{fitManyGPs}
\title{A wrapper to fit GPs on all columns of a matrix}
\usage{
fitManyGPs(mat, coord, ...)
}
\arguments{
\item{mat}{Matrix of observations}

\item{...}{passed onto \link{fitGPs}}
}
\value{
Matrix of fitted GP components
}
\description{
A wrapper to fit GPs on all columns of a matrix
}
