% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMsMulti.R
\name{GAMsMulti}
\alias{GAMsMulti}
\title{Fit GAMs and find correlations and standard error for data lists}
\usage{
GAMsMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families,
  n_points_grid,
  verbose,
  includeGPsmooth,
  findVariances = FALSE
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families, n_points_grid, includeGPsmooth}{See \link{GAMsSingle}}

\item{verbose}{Should info on type of analysis be printed?}

\item{findVariances}{Should variances be calculated? For internal use only}
}
\value{
A list named like Xl, containing all results
}
\description{
Wraps \link{GAMsSingle} for lists
}
