% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSigmas.R
\name{GaussKernel}
\alias{GaussKernel}
\title{Construct a covariance matrix using the Gaussian covariance kernel}
\usage{
GaussKernel(distMat, range)
}
\arguments{
\item{distMat}{The distance matrix of Cx and Ey}
}
\value{
The covariance matrix
}
\description{
Construct a covariance matrix using the Gaussian covariance kernel
}
