% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovPredGam.R
\name{getApproxVar}
\alias{getApproxVar}
\title{Get approximate variance of the}
\usage{
getApproxVar(vcovMat, cen, x, link)
}
\arguments{
\item{vcovMat}{Variance-covariance matrix of spline predictions}

\item{cen}{Centered observations}

\item{x}{Observations}

\item{link}{Link function}
}
\value{
Variance-covariance matrix of the spline predictions
}
\description{
Get approximate variance of the
}
