% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSize}
\alias{getSize}
\title{Construct the size variable}
\usage{
getSize(X, Y, normX, normY, size, scaleBySampleSums)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{normX, normY}{Character strings, indicating what normalization is required
for X and Y matrices, respectively, before plotting, see details.}

\item{size}{The maximum output size}

\item{scaleBySampleSums}{A boolean, should the size of the spots be scaled by their sample sum, e.g. library size or total ion count? Recommended to reflect differences in certainty depending on sample sums.}
}
\value{
The final point size
}
\description{
Construct the size variable
}
