% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logNorm}
\alias{logNorm}
\title{Log normalize a data matrix}
\usage{
logNorm(x, pseudoCount = 1e-08)
}
\arguments{
\item{x}{The matrix}

\item{pseudoCount}{A pseudocount added to avoid taking the log of zero}
}
\value{
A log-normalized matrix
}
\description{
Normalize to relative expression, add pseudocount and log-normalize
}
\examples{
mat = matrix(rpois(2000), 40, 50)
lnMat = logNorm(mat)
}
