% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapModTtest.R
\name{matchCoords}
\alias{matchCoords}
\title{Match coordinates to the nearest neighbour}
\usage{
matchCoords(Cx, Ey, mapToFinest)
}
\arguments{
\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{mapToFinest}{Passed onto \link{wrapModTtest}}
}
\value{
A list with components
\item{coordMat}{The new, single coordinate matrix}
\item{matToX}{A boolean, should X coordinates be taken}
\item{id}{The index of the coordinates to retain}
}
\description{
Match coordinates to the nearest neighbour
}
