% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapMoransI.R
\name{matheronVariograms}
\alias{matheronVariograms}
\title{Estimate variograms using Matheron's binning estimator for many features at once}
\usage{
matheronVariograms(X, Cx, width, cutoff)
}
\arguments{
\item{X}{Outcome matrix}

\item{Cx}{Coordinate matrix}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{variogram}}
}
\value{
A list of variograms
}
\description{
Estimate variograms using Matheron's binning estimator for many features at once
}
\note{
Only the Gaussian variogram (model = "Gau") is implemented here
}
