% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransISingle.R
\name{matheronVariograms}
\alias{matheronVariograms}
\title{Estimate variograms using Matheron's binning estimator for many features at once, and evaluate}
\usage{
matheronVariograms(X, Cx, width, cutoff, variogramModels)
}
\arguments{
\item{X}{Outcome matrix}

\item{Cx}{Coordinate matrix}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{variogramModels}{A character string, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}
}
\value{
A list of evaluated variograms
}
\description{
Estimate variograms using Matheron's binning estimator for many features at once, and evaluate
}
\details{
The best fitting variogram model, measured by the squared error, will be used.
}
