% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapMoransI.R
\name{matheronVariograms}
\alias{matheronVariograms}
\title{Estimate variograms using Matheron's binning estimator for many features at once}
\usage{
matheronVariograms(X, Cx, width, cutoff, model)
}
\arguments{
\item{cutoff, width}{passed onto \link[gstat]{variogram}}
}
\description{
Estimate variograms using Matheron's binning estimator for many features at once
}
