% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransIMulti.R
\name{MoransIMulti}
\alias{MoransIMulti}
\title{Find all Moran's I statistics for a list of matrices}
\usage{
MoransIMulti(Xl, Yl, Cxl, Eyl, verbose, ...)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{verbose}{Should info on type of analysis be printed?}

\item{...}{passed onto \link{MoransISingle}}
}
\value{
A list of named Moran's I vectors
}
\description{
Find all raw Moran's I values between lists of observations matrices from different modalities.
Observation matrices are scaled and centered prior to calculation
}
\seealso{
\link{buildWeightMat}
}
