% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransIMulti.R
\name{MoransIMulti}
\alias{MoransIMulti}
\title{Find all Moran's I statistics for a list of matrices}
\usage{
MoransIMulti(Xl, Yl, Cxl, Eyl, findVariances, verbose, ...)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{findVariances}{Should variances be calculated? For internal use only, not default for multi-image analysis}

\item{verbose}{Should info on type of analysis be printed?}

\item{...}{passed onto \link{MoransISingle}}
}
\value{
A list of Moran's I estimates, standard errors and maximum values
}
\description{
Find all Moran's I values for lists of observations matrices from different modalities,
by caling the \link{MoransISingle} function.
}
\seealso{
\link{MoransISingle}
}
