% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransISingle.R
\name{MoransISingle}
\alias{MoransISingle}
\title{Calculate bivariate Moran's I between two modality matrix, with variance and p-value}
\usage{
MoransISingle(
  X,
  Y,
  Cx,
  Ey,
  wo,
  etas,
  numNN,
  cutoff,
  width,
  verbose,
  findMaxW,
  variogramModels,
  returnVarsMoransI,
  ...
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{wo, numNN}{passed onto \link{buildWeightMat}}

\item{etas}{A vector of decay parameters for the weight function, see details}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{verbose}{Should info on type of analysis be printed?}

\item{findMaxW}{Is the maximum bivariate Moran's I needed?}

\item{variogramModels}{A character string, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{returnVarsMoransI}{A boolean, are variances of Moran's I to be returned?}

\item{...}{passed onto \link[gstat]{variogram}}
}
\value{
A dataframe of results sorted by p-value, also containing the estimated Moran's I statistic and its variance.
In addition, the maximum value of the Moran's I statistic
}
\description{
The variance calculation requires estimation of the spatial autocorrelation structure of every feature separately
}
\details{
By default, a number of range parameters and corresponding weight matrices are screened for spatial association,
and their p-value combined using the Cauchy combination rule by \insertCite{Liu2020}{sbivar}.
The maximum value of the bivariate Moran's I statistics are returned conditionally,
as it is computation intensive and not always needed.
}
\references{
\insertAllCited{}
}
