% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransISingle.R
\name{MoransISingle}
\alias{MoransISingle}
\title{Calculate bivariate Moran's I between two modality matrix, with variance and p-value}
\usage{
MoransISingle(
  X,
  Y,
  Cx,
  Ey,
  wo,
  etas,
  numNNs,
  cutoff,
  width,
  verbose,
  findMaxW,
  variogramModels,
  returnSEsMoransI,
  featuresX,
  featuresY,
  findVariances = TRUE,
  ...
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{wo}{type of weight parameter, passed onto \link{buildWeightMat}}

\item{numNNs, etas}{Vectors of weight matrix parameters, whose elements are passed onto \link{buildWeightMat}}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{verbose}{Should info on type of analysis be printed?}

\item{findMaxW}{Is the maximum bivariate Moran's I needed?}

\item{variogramModels}{A character vector, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{returnSEsMoransI}{A boolean, are standard errors of Moran's I to be returned?}

\item{featuresX, featuresY}{Features to be tested. Setting these allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization.}

\item{findVariances}{Should variances be calculated? For internal use only, not default for multi-image analysis}

\item{...}{passed onto \link[gstat]{variogram}}
}
\value{
A dataframe of results sorted by p-value, also containing the estimated Moran's I statistic and its variance.
In addition, the maximum value of the Moran's I statistic, and the parameters of the weight matrix
}
\description{
The variance calculation requires estimation of the spatial autocorrelation structure of every feature separately
}
\details{
By default, a number of range parameters and corresponding weight matrices are screened for spatial association,
and their p-value combined using the Cauchy combination rule by \insertCite{Liu2020}{sbivar}.
The maximum value of the bivariate Moran's I statistics are returned conditionally,
as it is computation intensive and not always needed.
}
\note{
No multithreading is implemented for the variance calculation, as the matrix calculations involved
may use inherent multithreading with OpenBLAS.
}
\references{
\insertAllCited{}
}
