% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransISinglePPP.R
\name{MoransISinglePPP}
\alias{MoransISinglePPP}
\title{Calculate bivariate Moran's I between a point pattern and a quantitative modality, with variance and p-value}
\usage{
MoransISinglePPP(
  X,
  Y,
  Ey,
  wo,
  etas,
  cutoff,
  width,
  verbose,
  variogramModels,
  returnSEsMoransI,
  featuresX,
  featuresY,
  findVariances = TRUE,
  ...
)
}
\arguments{
\item{X}{A marked point pattern of class \link[spatstat.geom]{ppp}}

\item{Y}{Matrix of omics measurements for the second, quantitative modality}

\item{Ey}{Coordinate matrix of dimension two, belonging to Y}

\item{wo}{type of weight parameter, passed onto \link{buildWeightMat}}

\item{etas}{Vectors of weight matrix parameters, whose elements are passed onto \link{buildWeightMat}}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{verbose}{Should info on type of analysis be printed?}

\item{variogramModels}{A character vector, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{returnSEsMoransI}{A boolean, are standard errors of Moran's I to be returned?}

\item{featuresX, featuresY}{Features to be tested. Defaults to all features, but specifying them allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization for Y.}

\item{findVariances}{Should variances be calculated? For internal use only}

\item{...}{passed onto \link[gstat]{variogram}}
}
\value{
A dataframe of results sorted by p-value, also containing the estimated Moran's I statistic and its variance.
In addition, the maximum value of the Moran's I statistic, and the parameters of the weight matrix
}
\description{
The variance calculation requires estimation of the spatial autocorrelation structure of every feature of the second modality, using Matheron's variogram estimator \insertCite{Matheron1963}{sbivar}.
}
\details{
By default, a number of range parameters and corresponding weight matrices are screened for spatial association,
and their p-value combined using the Cauchy combination rule by \insertCite{Liu2020}{sbivar}.
The maximum value of the bivariate Moran's I statistics are different for every gene pair unlike for \link{MoransISingle} as the weight matrix is random. As this presents too much computation, no maximum values are calculated.
}
\note{
No multithreading is implemented for the variance calculation, as the matrix calculations involved
may use inherent multithreading with OpenBLAS.
}
\references{
\insertAllCited{}
}
