% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R, R/allGenerics.R
\name{plotGAMs}
\alias{plotGAMs}
\alias{plotGAMsTopPair}
\title{Plot the fitted splines, and the correlation between them}
\usage{
plotGAMs(
  X,
  Y,
  Cx,
  Ey,
  features,
  scaleFun = "scaleMinusOne",
  families = list(X = gaussian(), Y = gaussian()),
  addTitle = TRUE,
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  n_points_grid = 600,
  Gamm = FALSE,
  correlation = corGaus(form = ~x + y, nugget = TRUE, value = c(1, 0.25)),
  ...
)

plotGAMsTopPair(x, topRank = 1, parameter = "Intercept", ...)
}
\arguments{
\item{X, Y}{Matrices of omics measurements, or lists thereof}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two, or lists thereof}

\item{features}{The features to plot}

\item{scaleFun}{The scaling function to be applied before plotting}

\item{families}{A vector of length 2 giving the distributional families
for the outcome values. See details of \link{sbivarSingle}.}

\item{addTitle}{A boolean, should a title be plotted}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{Gamm}{A boolean, should a gamm with Gaussian covariance kernel be fitted? See \link[mgcv]{gamm}}

\item{correlation}{Correlation structure, passed onto \link{fitGP} or \link{fitGAM}}

\item{...}{passed onto \link{fitGAM}}

\item{results}{Result of a call to \link{sbivar} (single-image) or to
\link{fitLinModels} (multi-image)}
}
\value{
A ggplot object
}
\description{
Spawns a three-panel plot with the splines fitted for the two variables, plus a visualization
of regions with positive and negative correlations between those splines.
The splines are refitted using \link{fitGAM}, so no GAM-objects can be provided.
This function takes entire outcome matrices X and Y as argument,
to be able to account for offsets in refitting the GAMs.
plotGAMsTopResults() plots the feature with the smallest p-value in the 'results' object.
}
\note{
Both spline surfaces are scaled to the [-1,1] range, the same as
the correlation has naturally, for legibility.
}
\examples{
# Single image
example(sbivar, "sbivar")
plotGAMs(X, Y, Cx, Ey, features = c("X1", "Y2"))
plotGAMsTopPair(resGAMs, X, Y, Cx = Cx, Ey = Ey)
# Multi image, arbitrary pair
data(Vicari)
plotGAMs(Vicari$TranscriptOutcomes, Vicari$MetaboliteOutcomes,
    Vicari$TranscriptCoords, Vicari$MetaboliteCoords,
    features = c("Pcp4", "Dopamine")
)
}
