% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{plotGAMs}
\alias{plotGAMs}
\alias{plotGAMsTopResults}
\title{Plot the fitted splines, and the correlation between them}
\usage{
plotGAMs(
  X,
  Y,
  Cx,
  Ey,
  features,
  offsets = list(),
  scaleFun = "scaleMinusOne",
  families = list(X = gaussian(), Y = gaussian()),
  addTitle = TRUE,
  n_points_grid = 600,
  multi = FALSE,
  ...
)

plotGAMsTopResults(
  results,
  X,
  Y,
  Cx,
  Ey,
  topRank = 1,
  parameter = "Intercept",
  ...
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{features}{The features to plot}

\item{offsets}{List of length two with offsets}

\item{scaleFun}{The scaling function to be applied before plotting}

\item{families}{A vector of length 2 giving the distributional families
for the outcome values. See details of \link{sbivarSingle}.}

\item{addTitle}{A boolean, should a title be plotted}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{multi}{Flag to indicate this is part of a multi-image analysis.
It only matters for the messages printed.}

\item{...}{passed onto \link{fitGAM}}

\item{results}{Results returned by \link{sbivarSingle}}

\item{topRank}{An integer, the feature pair with the rank-th smallest p-value is plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.
The default is the intercept, i.e. the overall effect.}

\item{x, y}{outcome vectors}

\item{newGrid}{The grid in which to evaluate the GAMs}

\item{resultsSingle}{Result of a call to \link{sbivarSingle}}
}
\value{
A ggplot object
}
\description{
Spawns a three-panel plot with the splines fitted for the two variables, plus a visualization
of regions with positive and negative correlations between those splines.
The splines are refitted using \link{fitGAM}, so no gam-objects can be provided.
plotGAMsFromMatrix() takes entire outcome matrices X and Y as argument,
to be able to account for offsets in refitting the GAMs.#'
plotGAMsTopResults() plots the feature with the smallest p-value in resultsSingle.
}
\note{
Both spline surfaces are scaled to the [-1,1] range, the same as
the correlation has naturally, for legibility.
}
\examples{
#Single image
example(fitLinModels, "sbivar")
plotGAMs(X, Y, Cx, Ey, features = c("X1", "Y3"))
plotGAMsTopResults(resGAMs, X, Y, Cx = Cx, Ey = Ey)
# Multi image
plotGAMsTopResults(resMoran, Vicari$TranscriptOutcomes, Vicari$MetaboliteOutcomes,
Vicari$TranscriptCoords, Vicari$MetaboliteCoords)
}
