% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{plotGAMs}
\alias{plotGAMs}
\alias{plotGAMsTopResults}
\title{Plot the fitted splines, and the correlation between them}
\usage{
plotGAMs(
  x,
  y,
  Cx,
  Ey,
  newGrid,
  offsets = list(),
  scaleFun = "scaleMinusOne",
  families = list(X = gaussian(), Y = gaussian()),
  addTitle = TRUE,
  n_points_grid = min(length(x), length(y)),
  features = c("x", "y"),
  ...
)

plotGAMsTopResults(resultsSingle, X, Y, Cx, Ey, ...)
}
\arguments{
\item{x, y}{outcome vetors}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{newGrid}{The grid in which to evaluate the GAMs}

\item{offsets}{List of length two with offsets}

\item{scaleFun}{The scaling function to be applied before plotting}

\item{families}{A vector of length 2 giving the distributional families for the outcome values. See details.}

\item{addTitle}{A booleam, should a title be plotted}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{features}{The features to plot}

\item{...}{passed onto \link{fitGAM}}

\item{resultsSingle}{Results returned by \link{sbivarSingle}}

\item{X, Y}{Matrices of omics measurements}
}
\value{
A ggplot object
}
\description{
Spawns a three-panel plot with the splines fitted for the two variables, plus a visualization
of regions with positive and negative correlations between those splines.
}
\note{
Both spline surfaces are scaled to the [-1,1] range,
the same as the correlation has naturally, for legibility.
}
\examples{
example(sbivarSingle, "sbivar")
plotGAMs(X[, 1], Y[, 1], Cx, Ey)
}
