% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{plotGAMs}
\alias{plotGAMs}
\title{Title}
\usage{
plotGAMs(
  x,
  y,
  Cx,
  Ey,
  newGrid,
  offsets,
  scaleFun = "scaleMinusOne",
  families = list(X = gaussian(), Y = gaussian()),
  addTitle = TRUE,
  ...
)
}
\arguments{
\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{newGrid}{The new grid of points in which to evaluate the GAMs}

\item{offsets}{List of length two with offsets}

\item{scaleFun}{The scaling function to be applied before plotting}

\item{families}{A vector of length 2 giving outcome values.}

\item{addTitle}{A booleam, should a title be plotted}

\item{...}{passed onto \link{fitGAM}}
}
\value{
A ggplot object
}
\description{
Title
}
\examples{
example(sbivar, "sbivar")
plot(X[, 1], Y[, 1], Cx, Ey)
}
