% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{plotGAMs}
\alias{plotGAMs}
\title{Plot the fitted splines, and the correlation between them}
\usage{
plotGAMs(
  x,
  y,
  Cx,
  Ey,
  newGrid,
  offsets = list(),
  scaleFun = "scaleMinusOne",
  families = list(X = gaussian(), Y = gaussian()),
  addTitle = TRUE,
  n_points_grid = 500,
  ...
)
}
\arguments{
\item{x, y}{outcome vetors}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{newGrid}{The grid in which to evaluate the GAMs}

\item{offsets}{List of length two with offsets}

\item{scaleFun}{The scaling function to be applied before plotting}

\item{families}{A vector of length 2 giving outcome values.}

\item{addTitle}{A booleam, should a title be plotted}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{...}{passed onto \link{fitGAM}}
}
\value{
A ggplot object
}
\description{
Plot the fitted splines, and the correlation between them
}
\examples{
example(sbivarSingle, "sbivar")
plotGAMs(X[, 1], Y[, 1], Cx, Ey)
}
