% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{plotGAMsFromMatrix}
\alias{plotGAMsFromMatrix}
\title{A wrapper frunction to plot from matrices}
\usage{
plotGAMsFromMatrix(X, Y, featx, featy, Cx, Ey, families, ...)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{featx, featy}{The features to plot}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{families}{A vector of length 2 giving outcome values.}

\item{...}{Passed onto \link{plotGAMs}.}
}
\description{
A wrapper frunction to plot from matrices
}
