% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAMs.R
\name{plotGAMsFromMatrix}
\alias{plotGAMsFromMatrix}
\title{A wrapper frunction to plot from matrices}
\usage{
plotGAMsFromMatrix(
  X,
  Y,
  features,
  Cx,
  Ey,
  families = list(X = gaussian(), Y = gaussian()),
  ...
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{families}{A vector of length 2 giving outcome values.}

\item{...}{Passed onto \link{plotGAMs}.}
}
\value{
See \link{plotGAMs}
}
\description{
A wrapper frunction to plot from matrices
}
\examples{
example(sbivarSingle, "sbivar")
plotGAMsFromMatrix(X, Y, featX = "X1", featY = "Y1", Cx = Cx, Ey = Ey)
}
