% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-plotGAMsTopPair.R
\name{plotGAMsTopPair}
\alias{plotGAMsTopPair}
\alias{plotGAMsTopPair,sbivarResultsGAMs-method}
\title{Plot GAMs of top features}
\usage{
plotGAMsTopPair(x, topRank = 1, parameter = "Intercept", ...)

\S4method{plotGAMsTopPair}{sbivarResultsGAMs}(x, topRank, parameter, X, Y, Cx, Ey, ...)
}
\description{
Fit GAMS for the top feature pair chosen, and plot the resulting spline surfaces
}
