% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-plotGAMsTopPair.R
\name{plotGAMsTopPair}
\alias{plotGAMsTopPair}
\alias{plotGAMsTopPair,sbivarResultsGAMs-method}
\title{Plot GAMs of top features}
\usage{
plotGAMsTopPair(x, X, Y, Cx, Ey, topRank = 1, parameter = "Intercept", ...)

\S4method{plotGAMsTopPair}{sbivarResultsGAMs}(x, X, Y, Cx, Ey, topRank = 1, parameter = "Intercept", ...)
}
\arguments{
\item{x}{a sbivarResultsGAMs object}

\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{topRank}{An integer, the feature pair with the rank-th smallest p-value is plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.
The default is the intercept, i.e. the overall effect.}

\item{...}{passed onto lower level functions}
}
\value{
a ggplot2 object
}
\description{
Fit GAMS for the top feature pair chosen, and plot the resulting spline surfaces
}
\examples{
example(sbivar, "sbivar")
plotGAMsTopPair(resGAMs, X, Y, Cx, Ey)
}
