% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPair.R
\name{plotPairSingle}
\alias{plotPairSingle}
\alias{plotPairMulti}
\alias{plotPairSingleVectors}
\title{Plot a specified feature pair for the data provided}
\usage{
plotPairSingle(
  X,
  Y,
  Cx,
  Ey,
  features,
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  assayX,
  assayY,
  scaleBySampleSums = FALSE,
  size = 1.5,
  ...
)

plotPairMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  features,
  normX = c("none", "rel", "log"),
  scaleBySampleSums = FALSE,
  normY = c("none", "rel", "log"),
  size = 1.25,
  assayX,
  assayY,
  theme = theme_bw()
)

plotPairSingleVectors(
  x,
  y,
  Cx,
  Ey,
  size,
  modalityNames = c("Modality X", "Modality Y"),
  theme = theme_bw(),
  ...
)
}
\arguments{
\item{X}{the input object, containing measurements of the first modality, see methods('sbivar')}

\item{Y}{Matrix or SpatialExperiment object of second modality}

\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{features}{Feature vector of length 2 to be plotted}

\item{normX, normY}{Character strings, indicating what normalization is required
for X and Y matrices, respectively, before plotting, see details.}

\item{assayX, assayY}{Assay names to be used in the analysis,
see \link[SummarizedExperiment]{assay}}

\item{scaleBySampleSums}{boolean, should the point size be scaled by sample sums?}

\item{size}{Point size}

\item{...}{additional constructor and analysis arguments}

\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{theme}{the ggplot2 theme}

\item{x, y}{Outcome vectors}

\item{modalityNames}{Names to be given to the modalities,
appearing in the strip text of the columns. For plotTopPair() and
plotPairSingle(), the feature names are used.}
}
\value{
a ggplot2 object
}
\description{
Plot a specified feature pair for the data provided
}
\examples{
example(sbivar, "sbivar")
plotPairSingle(X, Y, Cx, Ey, features = c("X1", "Y1"))
}
