% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPair.R
\name{plotPairSingle}
\alias{plotPairSingle}
\title{Plot a feature pair}
\usage{
plotPairSingle(x, y, Cx, Ey)
}
\arguments{
\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{resultsSingle, resultsMulti}{Results, returned by \link{sbivarSingle} and
\link{extractResultsNulti}, respectively.}
}
\value{
A ggplot2 object
}
\description{
Plot a chosen feature pair for a single or list of matrices,
or plot the highest ranking feature pair.
}
