% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairMulti.R
\name{plotPairMulti}
\alias{plotPairMulti}
\alias{plotTopResultsMulti}
\title{Plot a chosen or top feature in the analysis for a series of images}
\usage{
plotPairMulti(Xl, Yl, Cxl, Eyl, featx, featy)

plotTopResultsMulti(resultsMulti, Xl, Yl, Cx, Ey, parameter = "Intercept")
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}
}
\description{
Plot a chosen or top feature in the analysis for a series of images
}
\examples{
plotPairMulti(Xl, Yl, Cxl, Eyl, featx = "X1", featy = "Y1")
plotTopResultsMulti(estCorrelations, Xl, Yl, Cxl, Eyl)
}
