% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPair.R, R/allGenerics.R,
%   R/methods-plotTopPair.R
\name{plotPairSingle}
\alias{plotPairSingle}
\alias{plotPairMulti}
\alias{plotPairSingleVectors}
\alias{plotTopPair}
\alias{plotTopPair,sbivarResults-method}
\title{Plot a the top feature pair according to a sbivar analysis}
\usage{
plotPairSingle(
  X,
  Y,
  Cx,
  Ey,
  features,
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  assayX,
  assayY,
  scaleBySampleSums = FALSE,
  size = 1.5,
  ...
)

plotPairMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  features,
  normX = c("none", "rel", "log"),
  scaleBySampleSums = FALSE,
  normY = c("none", "rel", "log"),
  size = 1.25,
  assayX,
  assayY,
  theme = theme_bw()
)

plotPairSingleVectors(
  x,
  y,
  Cx,
  Ey,
  size,
  modalityNames = c("Modality X", "Modality Y"),
  theme = theme_bw(),
  ...
)

plotTopPair(
  x,
  ...,
  topRank = 1,
  parameter = "Intercept",
  scaleBySampleSums = FALSE,
  normX = x@normX,
  normY = x@normY
)

\S4method{plotTopPair}{sbivarResults}(
  x,
  ...,
  topRank = 1,
  parameter = "Intercept",
  scaleBySampleSums = FALSE,
  normX = x@normX,
  normY = x@normY
)
}
\arguments{
\item{X}{the input object, containing measurements of the first modality, see methods('sbivar')}

\item{Y}{Matrix or SpatialExperiment object of second modality}

\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{features}{Feature vector of length 2 to be plotted}

\item{normX, normY}{Character strings, indicating what normalization is required
for X and Y matrices, respectively, before plotting, see details.}

\item{assayX, assayY}{Assay names to be used in the analysis,
see \link[SummarizedExperiment]{assay}}

\item{scaleBySampleSums}{A boolean, should the size of the spots be scaled by their sample sum, e.g. library size or total ion count? Recommended to reflect differences in certainty depending on sample sums.}

\item{size}{Point size}

\item{...}{passed onto lower level functions}

\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{theme}{the ggplot2 theme}

\item{x, y}{Outcome vectors}

\item{modalityNames}{Names to be given to the modalities,
appearing in the strip text of the columns. For plotTopPair() and
plotPairSingle(), the feature names are used.}

\item{topRank}{An integer, the feature pair with the rank-th smallest p-value is plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.
The default is the intercept, i.e. the overall effect.}

\item{results}{Results returned by \link{sbivarSingle} or \link{extractResultsMulti}}
}
\value{
A ggplot object
}
\description{
Plot a chosen feature pair, or the highest ranking feature pair,
for a single image or multiple images.
}
\details{
For sequence count data, such as transcriptomics, normalization
may be indicated to achieve clear plots (normX = "rel" or "log", see  \link{normMat}).
The normalization used for plotting is not necessarily the same as the one used for the analysis.
}
\examples{
### Single image
# Single image analysis on synthetic data
n <- 8e1
m <- 9e1
p <- 3
k <- 2
X <- matrix(rnorm(n * p), n, p,
    dimnames =
        list(paste0("sampleX", seq_len(n)), paste0("X", seq_len(p)))
)
Y <- matrix(rnorm(m * k), m, k,
    dimnames =
        list(paste0("sampleY", seq_len(m)), paste0("Y", seq_len(k)))
)
Cx <- matrix(runif(n * 2), n, 2, dimnames = list(rownames(X), c("x", "y")))
Ey <- matrix(runif(m * 2), m, 2, dimnames = list(rownames(Y), c("x", "y")))
resMoransI <- sbivar(X, Y, Cx, Ey, method = "Moran's I")
# Plot the feature pair with the most significant signal
plotTopPair(resMoransI, X, Y, Cx, Ey)
# Plot an arbitrary feature pair
plotPairSingle(X, Y, Cx, Ey, features = c("X1", "Y1"))
### Multi image
data(Vicari)
# Plot an arbitrary feature pair
plotPairMulti(Vicari$TranscriptOutcomes, Vicari$MetaboliteOutcomes,
    Vicari$TranscriptCoords, Vicari$MetaboliteCoords,
    normX = "rel", normY = "rel", features = c("Gnas", "Tocopherol")
)
}
\seealso{
\link{extractResultsMulti}, \link{fitLinModels}
}
