% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-plotTopPair.R
\name{plotTopPair}
\alias{plotTopPair}
\alias{plotTopPair,sbivarResults-method}
\title{Plot a the top feature pair according to a sbivar analysis}
\usage{
plotTopPair(
  x,
  ...,
  topRank = 1,
  parameter = "Intercept",
  scaleBySampleSums = FALSE,
  normX = x@normX,
  normY = x@normY
)

\S4method{plotTopPair}{sbivarResults}(
  x,
  ...,
  topRank = 1,
  parameter = "Intercept",
  scaleBySampleSums = FALSE,
  normX = x@normX,
  normY = x@normY
)
}
\arguments{
\item{...}{passed onto lower level functions}

\item{topRank}{An integer, the feature pair with the rank-th smallest p-value is plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.
The default is the intercept, i.e. the overall effect.}

\item{scaleBySampleSums}{A boolean, should the size of the spots be scaled by their sample sum, e.g. library size or total ion count? Recommended to reflect differences in certainty depending on sample sums.}
}
\value{
A ggplot object
}
\description{
Plot a chosen feature pair, or the highest ranking feature pair,
for a single image or multiple images.
}
\details{
For sequence count data, such as transcriptomics, normalization
may be indicated to achieve clear plots (normX = "rel" or "log", see  \link{normMat}).
The normalization used for plotting is not necessarily the same as the one used for the analysis.
}
\examples{
### Single image
# Single image analysis on synthetic data
n <- 8e1
m <- 9e1
p <- 3
k <- 2
X <- matrix(rnorm(n * p), n, p,
    dimnames =
        list(paste0("sampleX", seq_len(n)), paste0("X", seq_len(p)))
)
Y <- matrix(rnorm(m * k), m, k,
    dimnames =
        list(paste0("sampleY", seq_len(m)), paste0("Y", seq_len(k)))
)
Cx <- matrix(runif(n * 2), n, 2, dimnames = list(rownames(X), c("x", "y")))
Ey <- matrix(runif(m * 2), m, 2, dimnames = list(rownames(Y), c("x", "y")))
resMoransI <- sbivar(X, Y, Cx, Ey, method = "Moran's I")
# Plot the feature pair with the most significant signal
plotTopPair(resMoransI, X, Y, Cx, Ey)
# Plot an arbitrary feature pair
plotPairSingle(X, Y, Cx, Ey, features = c("X1", "Y1"))
### Multi image
data(Vicari)
# Plot an arbitrary feature pair
plotPairMulti(Vicari$TranscriptOutcomes, Vicari$MetaboliteOutcomes,
    Vicari$TranscriptCoords, Vicari$MetaboliteCoords,
    normX = "rel", normY = "rel", features = c("Gnas", "Tocopherol")
)
}
\seealso{
\link{extractResultsMulti}, \link{fitLinModels}
}
