% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPair.R
\name{plotTopPair}
\alias{plotTopPair}
\alias{plotPairSingle}
\alias{plotPairMulti}
\alias{plotPairSingleVectors}
\title{Plot a top feature pair from the analysis, or an arbitrary pair for a series of images}
\usage{
plotTopPair(results, rank = 1, parameter = "Intercept", ...)

plotPairSingle(
  X,
  Y,
  Cx,
  Ey,
  features,
  normalizationX = c("none", "log"),
  normalizationY = c("none", "log"),
  ...
)

plotPairMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  features,
  normalizationX = c("none", "log"),
  normalizationY = c("none", "log"),
  size = 1.25,
  assayX,
  assayY
)

plotPairSingleVectors(
  x,
  y,
  Cx,
  Ey,
  size = 1.25,
  modalityNames = c("Modality X", "Modality Y")
)
}
\arguments{
\item{results}{Results returned by \link{sbivarSingle}}

\item{rank}{An integer, the feature pair with the rank-th smallest p-value is plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.
The default is the intercept, i.e. the overall effect.}

\item{...}{passed onto lower level functions}

\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{features}{Feature vector of length 2 to be plotted}

\item{normalizationX, normalizationY}{Character strings, indicating what normalization is required
for X and Y matrices, respectively, before plotting, see details.}

\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{size}{Point size}

\item{x, y}{Outcome vectors}

\item{modalityNames}{Names to be given to the modalities,
appearing in the strip text of the columns. For plotTopPair() and
plotPairSingle(), the feature names are used.}
}
\value{
A ggplot object
}
\description{
Plot a chosen feature pair, or the highest ranking feature pair,
for a single image or multiple images.
}
\details{
For sequence count data, such as transcriptomics, log-normalization
may be indicated to achieve clear plots (normalizationX = "log", see  \link{logNorm}). The normalization used for plotting is
 not necessarily the same as the one used for the analysis.
}
\examples{
### Single image
example(sbivarSingle, "sbivar")
# Plot the feature pair with the most significant signal
plotTopPair(resModtTest, X, Y, Cx, Ey)
# Plot an arbitrary feature pair
plotPairSingle(X, Y, Cx, Ey, features = c("X1", "Y1"))
### Multi image
example(fitLinModels, "sbivar")
# Plot the feature pair with the most significant signal for a certain parameter,
#here "cofactor"
plotTopPair(resGams, Xl, Yl, Cxl, Eyl, parameter = "cofactor")
# Plot an arbitrary feature pair
plotPairMulti(Xl, Yl, Cxl, Eyl, features = c("X1", "Y1"))
}
\seealso{
\link{extractResultsMulti}, \link{fitLinModels}
}
