% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairPPP.R
\name{plotTopPairPPP}
\alias{plotTopPairPPP}
\alias{plotPairPPP}
\alias{plotPairSinglePPvec}
\title{Plot a single molecule point pattern and a quantitative outcome}
\usage{
plotTopPairPPP(results, X, Y, Ey, topRank = 1, ...)

plotPairPPP(X, Y, Ey, features, normY = "none", ...)

plotPairSinglePPvec(
  Cx,
  y,
  Ey,
  sizeX = 0.005,
  sizeY = 0.01,
  sideBySide = TRUE,
  modalityNames = c("Modality X", "Modality Y"),
  theme = theme_bw(),
  ...
)
}
\arguments{
\item{results}{Results returned by \link{sbivarSingle} or \link{extractResultsMulti}}

\item{X}{A point pattern of class ppp}

\item{Y}{Matrix or SpatialExperiment object of second modality}

\item{topRank}{An integer, the feature pair with the rank-th smallest p-value is plotted}

\item{...}{passed onto lower level functions}

\item{features}{Feature vector of length 2 to be plotted}

\item{Cx, Ey}{Coordinate matrices of dimension two, belonging to X and Y respectively}

\item{sizeX, sizeY}{Desired point sizes for corresponding modalities}

\item{sideBySide}{Boolean indicating the patterns are to be plotted side by side, or on top of each other if FALSE}

\item{modalityNames}{Names to be given to the modalities,
appearing in the strip text of the columns. For plotTopPairPPP() and
plotPairSinglePPP(), the feature names are used.}

\item{theme}{the ggplot2 theme}
}
\value{
A ggplot object
}
\description{
Plot a chosen feature pair consisting of a point pattern and a fixed-grid quantitative outcome, or the highest ranking feature pair,
for a single image
}
\note{
The point pattern is coloured blue, the value corresponding to 1
}
\examples{
n <- 8e1
m <- 9e1
p <- 3
k <- 2
Y <- matrix(rnorm(m * k), m, k,
    dimnames =
        list(paste0("sampleY", seq_len(m)), paste0("Y", seq_len(k)))
)
Ey <- matrix(runif(m * 2), m, 2, dimnames = list(rownames(Y), c("x", "y")))
# Single image analysis on synthetic point pattern + quantitative modality
library(spatstat.random)
lambda <- 8e1
p <- 3
PPP <- rmpoispp(lambda, types = paste0("X", seq_len(p)), win = owin(c(0, 1), c(0, 1)))
marks(PPP, drop = FALSE) <- data.frame(feature = marks(PPP)) # Make sure marks is a dataframe
resMoransIppp <- sbivar(X = PPP, Y, Ey, method = "Moran's I")
plotTopPairPPP(resMoransIppp, X = PPP, Y = Y, Ey = Ey)
# For overlay, do:
plotPairPPP(X = PPP, Y = Y, Ey = Ey, features = c("X1", "Y1"), sideBySide = FALSE)
}
\seealso{
\link{plotTopPair}, \link{sbivarSinglePPP}
}
