% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairSingle.R
\name{plotTopResultsSingle}
\alias{plotTopResultsSingle}
\alias{plotPairSingleVectors}
\title{Plot a feature pair}
\usage{
plotTopResultsSingle(resultsSingle, X, Y, Cx, Ey, ...)

plotPairSingleVectors(
  x,
  y,
  Cx,
  Ey,
  size = 2,
  modalityNames = c("feature 1", "feature 2")
)
}
\arguments{
\item{resultsSingle, resultsMulti}{Results, returned by \link{sbivarSingle} and
\link{extractResultsMulti}, respectively.}

\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{...}{Passed onto \link{plotPairSingleVectors}}

\item{x, y}{Outcome vectors}

\item{size}{Point size}

\item{featx, featy}{The features in X and Y to be plotted}

\item{parameter}{The linear model parameter used to find the feature with the strongest effect.}

\item{normalizationX, normalizationY}{A character string, indicating what normalization is required
for X and Y matrrices, respectively, before plotting.
This is not the same normalization as used for the analysis, see details.}
}
\value{
A ggplot2 object
}
\description{
Plot a chosen feature pair, or the highest ranking feature pair, for a singlepair of images,
}
\details{
For sequence count data, such as transcriptomics, log-normalization
seems indicated to achieve clear plots.
}
\examples{
example(sbivarSingle, "sbivar")
plotPairSingle(X, Y, Cx, Ey, features = c("X1", "Y1"))
plotTopResultsSingle(resModtTest, X, Y, Cx, Ey)
}
