% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairSingle.R
\name{plotTopResultsSingle}
\alias{plotTopResultsSingle}
\alias{plotPairSingle}
\alias{plotPairSingleVectors}
\title{Plot a feature pair}
\usage{
plotTopResultsSingle(resultsSingle, X, Y, Cx, Ey, ...)

plotPairSingle(
  X,
  Y,
  Cx,
  Ey,
  features,
  normalizationX = c("none", "log"),
  normalizationY = c("none", "log"),
  ...
)

plotPairSingleVectors(
  x,
  y,
  Cx,
  Ey,
  size = 1.25,
  modalityNames = c("Modality X", "Modality Y")
)
}
\arguments{
\item{resultsSingle}{Results returned by \link{sbivarSingle}}

\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{...}{Passed onto \link{plotPairSingleVectors}}

\item{features}{Feature vector of length 2 to be plotted}

\item{normalizationX, normalizationY}{Character strings, indicating what normalization is required
for X and Y matrices, respectively, before plotting, see details.}

\item{x, y}{Outcome vectors}

\item{size}{Point size}

\item{modalityNames}{Names to be given to the modalities,
appearing in the strip text of the columns. For plotTopResultsSingle() and
plotPairSingle(), the feature names are used.}
}
\value{
A ggplot2 object
}
\description{
Plot a chosen feature pair, or the highest ranking feature pair, for a single pair of images
}
\details{
For sequence count data, such as transcriptomics, log-normalization
seems indicated to achieve clear plots (normalizationX = "log", see  \link{logNorm}). The normalization used for plotting is
 not necessarily the same as the one used for the analysis.
}
\examples{
example(sbivarSingle, "sbivar")
# Plot the feature pair with the most significant signal
plotTopResultsSingle(resModtTest, X, Y, Cx, Ey)
# Plot an arbitrary feature pair
plotPairSingle(X, Y, Cx, Ey, features = c("X1", "Y1"))
}
