% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-sbivar.R,
%   R/sbivar-package.R
\docType{package}
\name{sbivar}
\alias{sbivar-package}
\alias{sbivar}
\alias{sbivar,MultiAssayExperiment-method}
\alias{sbivar,matrix-method}
\alias{sbivar,list-method}
\alias{sbivar,SpatialExperiment-method}
\title{Perform a spatial bivariate association analysis}
\usage{
sbivar(X, ...)

\S4method{sbivar}{MultiAssayExperiment}(X, ...)

\S4method{sbivar}{matrix}(X, Y, Cx, Ey, ...)

\S4method{sbivar}{list}(X, ...)

\S4method{sbivar}{SpatialExperiment}(X, Y, assayX, assayY, ...)
}
\arguments{
\item{X}{the input object, see methods('sbivar')}

\item{...}{additional constructor arguments}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{coordVars}{Names of coordinates}

\item{imageIdentifier}{A character vector of variables whose unique combinations
define the separate point patterns (images)}

\item{imageVars}{Covariates belonging to the point patterns}

\item{pointVars}{Names of event-wise covariates such as gene or cell for each single point}

\item{assaX, assayY}{Assay names to be used in the analysis}
}
\value{
The analysis result
}
\description{
Perform the spatial analysis of choice, either on a single or multiple imgaes

The sbivar package implements a suite of tests for Spatial BIVARiate association across omics modalities, with possibly disjoint coordinate sets. Implemented tests are generalized additive models (GAMs), modified t-test, bivariate Moran's I and Gaussian processes (GPs). Both single images and replicated experiments can be analysed.
}
\examples{
data(Vicari)
sbivarSingle(singleStx, singleMet, singleStxCoords, singleMetCoords,
method = "GAMs",  families = FamiliesViari)
}
\seealso{
\link[spatstat.geom]{hyperframe}

Useful links:
\itemize{
  \item \url{https://github.com/bivar/smoppix}
  \item Report bugs at \url{https://github.com/sthawinke/bivar/issues}
}

}
\author{
\strong{Maintainer}: Stijn Hawinkel \email{stijn.hawinkel@psb.ugent.be} (\href{https://orcid.org/0000-0002-4501-5180}{ORCID})

}
\keyword{internal}
