% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivar.R
\name{sbivar}
\alias{sbivar}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivar(
  X,
  Y,
  Cx,
  Ey,
  method = c("GAMs", "Modified t-test", "GPs"),
  wMat,
  wo = "distance",
  numNN = 8,
  n_points_grid = 500
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{method}{A character string, indicating which method to apply}

\item{wMat}{Optional, a weight matrix for calculating Moran's I}

\item{wo, numNN}{Passed on to \link{buildWeightMat}.}

\item{n_points_grid}{}
}
\value{
A list containing at least an entry "result", which contains p-values,
adjusted p-values and measures of association, sorted by increasing p-value.
}
\description{
Apply one or more tests for bivariate spatial association
}
