% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivar-package.R
\docType{package}
\name{sbivar}
\alias{sbivar-package}
\alias{sbivar}
\title{sbivar: Test for Spatial Bivariate Association Across Omics Types}
\description{
The sbivar package implements a suite of tests for Spatial BIVARiate association across omics modalities, with possibly disjoint coordinate sets. Implemented tests are generalized additive models (GAMs), modified t-test, bivariate Moran's I and Gaussian processes (GPs). Both single images and replicated experiments can be analysed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bivar/smoppix}
  \item Report bugs at \url{https://github.com/sthawinke/bivar/issues}
}

}
\author{
\strong{Maintainer}: Stijn Hawinkel \email{stijn.hawinkel@psb.ugent.be} (\href{https://orcid.org/0000-0002-4501-5180}{ORCID})

}
\keyword{internal}
