% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-sbivar.R,
%   R/sbivar-package.R
\docType{package}
\name{sbivar}
\alias{sbivar-package}
\alias{sbivar}
\alias{sbivar,matrix-method}
\alias{sbivar,list-method}
\alias{sbivar,SpatialExperiment-method}
\alias{sbivar,MultiAssayExperiment-method}
\title{Perform a spatial bivariate association analysis}
\usage{
sbivar(X, ...)

\S4method{sbivar}{matrix}(X, Y, Cx, Ey, ...)

\S4method{sbivar}{list}(X, Y, Cx, Ey, assayX = NULL, assayY = NULL, ...)

\S4method{sbivar}{SpatialExperiment}(X, Y, assayX, assayY, sample_id_x, sample_id_y = sample_id_x, ...)

\S4method{sbivar}{MultiAssayExperiment}(X, experimentX, experimentY, assayX, assayY, ...)
}
\arguments{
\item{X}{the input object, see methods('sbivar')}

\item{...}{additional constructor and analysis arguments}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{assayX, assayY}{Assay names to be used in the analysis,
see \link[SummarizedExperiment]{assay}}

\item{sample_id_x, sample_id_y}{If provided, these are used to discriminate between different images
included in the same SpatialExperiment objects X and Y. By default, they are assumed to be the same for both X and Y}

\item{experimentX, experimentY}{Names of the experiments in X to be used in the analysis}
}
\value{
The analysis result
}
\description{
Perform the spatial analysis of choice, either on a single or multiple images.
Depending on the input, the workhorse functions \link{sbivarSingle}
(for matrix, SpatialExperiment or MultiAssayExperiment) or \link{sbivarMulti} (for a list) are called.

The sbivar package implements a suite of tests for Spatial BIVARiate association across omics modalities, with possibly disjoint coordinate sets. Implemented tests are generalized additive models (GAMs), modified t-test, bivariate Moran's I and Gaussian processes (GPs). Both single images and replicated experiments can be analysed.
}
\examples{
# Single image analysis on synthetic data
n=7e1;m=9e1;p=3;k=4
X = matrix(rnorm(n*p), n, p, dimnames = list(NULL, paste0("X", seq_len(p))))
Y = matrix(rnorm(m*k), m, k, dimnames = list(NULL, paste0("Y", seq_len(k))))
Cx = matrix(runif(n*2), n, 2)
Ey = matrix(runif(m*2), m, 2)
colnames(Cx) = colnames(Ey) = c("x", "y")
resGAMs = sbivar(X, Y, Cx, Ey, method = "GAMs")
resModtTest = sbivar(X, Y, Cx, Ey, method = "Modified")
resModtTestJoint = sbivar(X, Y[seq_len(nrow(X)),], Cx, method = "Modified")
# Single image analysis on synthetic data, converted to SpatialExperiment
library(SpatialExperiment)
seX = SpatialExperiment(assays = list("transcripts" = t(X)), spatialCoords = Cx)
seY = SpatialExperiment(assays = list("metabolites" = t(Y)), spatialCoords = Ey)
resModtGPs = sbivar(seX, seY, assayX = "transcripts", assayY = "metabolites", method = "GPs")
#Multi-image analysis on Vicari data
data(Vicari)
sbivar(singleStx, singleMet, singleStxCoords, singleMetCoords,
method = "GAMs", families = list("X" = mgcv::nb(), "Y" = Gamma(lin = "log")))
}
\seealso{
\link{sbivarSingle}, \link{sbivarMulti}

Useful links:
\itemize{
  \item \url{https://github.com/sbivar/smoppix}
  \item Report bugs at \url{https://github.com/sthawinke/sbivar/issues}
}

}
\author{
\strong{Maintainer}: Stijn Hawinkel \email{stijn.hawinkel@psb.ugent.be} (\href{https://orcid.org/0000-0002-4501-5180}{ORCID})

}
\keyword{internal}
