% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarMulti.R
\name{sbivarMulti}
\alias{sbivarMulti}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivarMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families = list(X = gaussian(), Y = gaussian()),
  method = c("Moran's I", "GAMs", "Correlation"),
  wo = c("Gauss", "nn"),
  numNNs = c(4, 8, 24),
  etas = c(2e-04, 0.002, 0.02),
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  returnSEsMoransI = TRUE,
  variogramModels = c("Exp", "Lin"),
  width = cutoff/15,
  cutoff = sqrt(2)/3,
  pseudoCount = 1e-08,
  n_points_grid = 600,
  verbose = TRUE
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families, n_points_grid}{Passed onto \link{GAMsMulti}}

\item{method}{A character string, indicating which method to apply}

\item{wo, numNNs, etas}{Passed onto \link{MoransIsingle}}

\item{normX, normY}{Character vectors indicating normalization, see \link{normMat}}

\item{returnSEsMoransI}{A boolean, are standard errors of Moran's I to be returned?}

\item{variogramModels}{A character string, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{pseudoCount}{A pseudocount added prior to log-normalization to avoid taking the log of zero}

\item{verbose}{Should info on type of analysis be printed?}
}
\value{
A list containing
\item{estimates}{The estimated measures of association}
\item{method}{The method used to find these estimates}
\item{multi}{TRUE, a flag for the type of analysis}
\item{normX,normY}{As provided}
\item{families,wo,numNNs,etas}{Optional, as provided}
}
\description{
This function calculates measures of spatial association for every image.
The resulting estimates can then be analysed further using the \link{fitLinModels} function.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\seealso{
\link{fitLinModels}, \link{MoransIMulti}, \link{correlationsMulti}, \link{GAMsMulti}
}
