% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarMulti.R
\name{sbivarMulti}
\alias{sbivarMulti}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivarMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families = list(X = gaussian(), Y = gaussian()),
  method = c("Moran's I", "GAMs", "Correlation"),
  etas = 2 * 10^c(-6, -5, -4, -3),
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  variogramModels = c("Exp", "Lin"),
  width = cutoff/15,
  cutoff = sqrt(2)/3,
  pseudoCount = 1e-08,
  wo = c("Gauss", "nn"),
  numNN = c(4, 8, 24),
  n_points_grid = 600,
  verbose = TRUE
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families, n_points_grid}{Passed onto \link{GAMsMulti}}

\item{method}{A character string, indicating which method to apply}

\item{etas}{A vector of decay parameters for the weight function, see details}

\item{normX, normY}{Character vectors indicating normalization, "log" means log-normalization of relative abundances}

\item{variogramModels}{A character string, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{pseudoCount}{A pseudocount added prior to log-normalization to avoid taking the log of zero}

\item{wo, numNN}{Passed onto \link{buildWeightMat}}

\item{verbose}{Should info on type of analysis be printed?}
}
\value{
A list containing
\item{estimates}{The estimated measures of association}
\item{method}{The method used to find these estimates}
\item{multi}{TRUE, a flag for the type of analysis}
}
\description{
This function calculates measures of spatial association for every image.
The resulting estimates can then be analysed further using the \link{fitLinModels} function.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\seealso{
\link{fitLinModels}
}
