% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarMulti.R
\name{sbivarMulti}
\alias{sbivarMulti}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivarMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families = list(X = gaussian(), Y = gaussian()),
  method = c("Moran's I", "GAMs", "Correlation"),
  mapToFinest = FALSE,
  wo = c("exp", "distance", "nn"),
  numNN = 8,
  n_points_grid = 600,
  verbose = TRUE
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families, n_points_grid}{Passed onto \link{wrapGAMsMulti}}

\item{method}{A character string, indicating which method to apply}

\item{wo, numNN}{Passed onto \link{buildWeightMat}}

\item{verbose}{Should info on type of analysis be printed?}
}
\value{
A list containing
\item{estimates}{The estimated measures of association}
\item{method}{The method used to find these estimates}
\item{multi}{TRUE, a flag for the type of analysis}
}
\description{
This function calculates measures of spatial association for every image.
The resulting estimates can then be analysed further using the \link{fitLinModels} function.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\seealso{
\link{fitLinModels}
}
