% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarMulti.R
\name{sbivarMulti}
\alias{sbivarMulti}
\title{Estimate measures of bivariate spatial association for multiple images}
\usage{
sbivarMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families = list(X = gaussian(), Y = gaussian()),
  method = c("Moran's I", "GAMs", "Correlation"),
  wo = c("Gauss", "nn"),
  numNNs = c(4, 8, 24),
  etas = c(5e-06, 2e-04, 0.02),
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  variogramModels = c("Exp", "Lin"),
  width = cutoff/15,
  cutoff = sqrt(2)/3,
  pseudoCount = 1e-08,
  n_points_grid = 600,
  verbose = TRUE,
  findVariances = FALSE,
  findMaxW = TRUE,
  includeGPsmooth = TRUE
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families, n_points_grid}{Passed onto \link{GAMsMulti}}

\item{method}{A character string, indicating which method to apply}

\item{wo, numNNs, etas}{Passed onto \link{MoransISingle}}

\item{normX, normY, pseudoCount}{Normalization parameters, passed onto \link{normMat}}

\item{variogramModels}{A character vector, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{verbose}{Should info on type of analysis be printed?}

\item{findVariances}{Should variances be calculated? For internal use only}

\item{findMaxW}{Is the maximum bivariate Moran's I needed?}

\item{includeGPsmooth}{Should a Gaussian random field smoother for stochastic neighbourhood similarity be included?}
}
\value{
A list containing
\item{estimates}{The estimated measures of association}
\item{multi}{TRUE, a flag for the type of analysis}
\item{normX,normY,method}{As provided}
\item{families,wo,wParams}{Optional, as provided. wParams are either etas or numNNs}
}
\description{
This function calculates measures of spatial association for every image.
The resulting estimates can then be analysed further using the \link{fitLinModels} function.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\seealso{
\link{fitLinModels}, \link{MoransIMulti}, \link{correlationsMulti}, \link{GAMsMulti}
}
