% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarMulti.R
\name{sbivarMulti}
\alias{sbivarMulti}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivarMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  method = c("GAMs", "Correlation", "GPs", "Moran's I"),
  wMat,
  wo = c("distance", "nn"),
  numNN = 8,
  n_points_grid = 500,
  families = list(X = gaussian(), Y = gaussian())
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{method}{A character string, indicating which method to apply}

\item{wMat}{Optional, a weight matrix for calculating Moran's I}

\item{wo, numNN}{Passed on to \link{buildWeightMat}.}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{families}{A vector of length 2 giving outcome values.}
}
\value{
A list containing at least an entry "result", which contains p-values,
adjusted p-values and measures of association, sorted by increasing p-value.
}
\description{
Apply one or more tests for bivariate spatial association
}
\examples{
n=1e2;m=2e2;p=10;k=12
X = matrix(rnorm(n*p), n, p, dimnames = list(NULL, paste0("X", seq_len(p))))
Y = matrix(rnorm(m*k), m, k, dimnames = list(NULL, paste0("Y", seq_len(k))))
Cx = matrix(runif(n*2), n, 2)
Ey = matrix(runif(m*2), m, 2)
colnames(Cx) = colnames(Ey) = c("x", "y")
resGAMs = sbivarMulti(X, Y, Cx, Ey, method = "GAMs")
resMoran = sbivarMulti(X, Y, Cx, Ey, method = "Moran")
}
