% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarMulti.R
\name{sbivarMulti}
\alias{sbivarMulti}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivarMulti(
  Xl,
  Yl,
  Cxl,
  Eyl,
  families = list(X = gaussian(), Y = gaussian()),
  method = c("GAMs", "Correlation", "Moran's I"),
  mapToFinest = FALSE,
  wo = c("distance", "nn"),
  numNN = 8,
  n_points_grid = 500
)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{families}{A vector of length 2 giving outcome values.}

\item{method}{A character string, indicating which method to apply}

\item{mapToFinest}{A boolean, should the one-to-one mapping for modified t-test
occur to the dataset with the best resolution?}

\item{wo}{The weighting option, either "nn" or "distance}

\item{numNN}{An integer, the number of neighbours}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}
}
\value{
A list containing
\item{estimates}{The estimated measures of association}
\item{method}{The method used to find these estimates}
}
\description{
This function calculates measures of spatial association for every image.
The resulting estimates can then be analysed further usign the \link{fitLinModels} function.
}
\examples{
n=1e2;m=8e1;p=3;k=4
ims = 6
X = lapply(selfName(seq_len(ims)), function(i){n = rpois(1, n)
 matrix(rnorm(n*p), n, p, dimnames = list(NULL, paste0("X", seq_len(p))))
})
Y = lapply(selfName(seq_len(ims)), function(i){m = rpois(1, m)
 matrix(rnorm(m*k), m, k, dimnames = list(NULL, paste0("Y", seq_len(k))))
})
Cx = lapply(X, function(x){n = nrow(x)
    matrix(runif(n*2), n, 2, dimnames = list(NULL, c("x", "y")))
})
Ey = lapply(Y, function(y){m = nrow(y)
    matrix(runif(m*2), m, 2, dimnames = list(NULL, c("x", "y")))
})
estGAMs = sbivarMulti(X, Y, Cx, Ey, method = "GAMs")
estMoran = sbivarMulti(X, Y, Cx, Ey, method = "Moran")
estCorrelations = sbivarMulti(X, Y, Cx, Ey, method = "Correlation")
}
\seealso{
[fitLinModels()]
}
