% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarSingle.R
\name{sbivarSingle}
\alias{sbivarSingle}
\title{Test for bivariate spatial association in a single image}
\usage{
sbivarSingle(
  X,
  Y,
  Cx,
  Ey,
  method = c("Moran's I", "GAMs", "Modified t-test", "GPs"),
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  etas = c(2e-06, 6e-05, 0.002),
  findMaxW = FALSE,
  families = list(X = gaussian(), Y = gaussian()),
  n_points_grid = 600,
  verbose = TRUE,
  variogramModels = c("Exp", "Lin"),
  width = cutoff/15,
  cutoff = sqrt(2)/3,
  wo = c("Gauss", "nn"),
  numNN = c(4, 8, 24),
  pseudoCount = 1e-08,
  returnSEsMoransI = FALSE,
  GPmethod = c("REML", "ML"),
  gpParams,
  Quants = c(0.005, 0.5),
  numLscAlts = 5,
  optControl = lmeControl(opt = "optim", maxIter = 500, msMaxIter = 500, niterEM = 1000,
    msMaxEval = 1000),
  corStruct = corGaus(form = ~x + y, nugget = TRUE, value = c(1, 0.25))
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{method}{A character string, indicating which method to apply}

\item{normX, normY}{Character vectors indicating normalization, "log" means log-normalization of relative abundances}

\item{etas}{A vector of decay parameters for the weight function, see details}

\item{findMaxW}{Is the maximum bivariate Moran's I needed?}

\item{n_points_grid, families}{Passed onto \link{GAMsSingle}}

\item{verbose}{Should info on type of analysis be printed?}

\item{variogramModels}{A character string, indicating the variogram model passed onto \link[gstat]{vgm}.
Currently, only "Exp" and "Lin" are implemented for computational reasons.}

\item{cutoff, width}{Cutoff and width of the variogram estimation, passed onto \link[gstat]{vgm}}

\item{wo, numNN}{passed onto \link{buildWeightMat}}

\item{pseudoCount}{A pseudocount added prior to log-normalization to avoid taking the log of zero}

\item{returnSEsMoransI}{A boolean, are standard errors of Moran's I to be returned?}

\item{GPmethod, Quants, numLscAlts, optControl, corStruct}{Passed onto \link{fitGP}}

\item{gpParams}{Parameters of the Gaussian processes, see details}
}
\value{
A list with at least the following components
\item{result}{A matrix which contains at least a p-values ("pVal") and a
Benjamini-Hochberg adjusted p-value ("pAdj"), sorted by increasing p-value.}
\item{families}{As provided}
\item{method}{As provided}
\item{multi}{FALSE, a flag for the type of analysis}
}
\description{
The tests can be applied to either disjoint or joint coordinate sets.
}
\details{
If only Cx is supplied and X and Y have the same number of rows, a joint analysis is performed
If Cx and Ey are provided, and X and Y have the same number of rows, equality of Cx and Ey is checked.
If true, a joint analysis is run, with a warning.

Any normalization of the data should happen prior to calling this function.
For instance, count data or metabolome data are best scaled to relative values and log-normalized prior to fitting GPs.
For GAMs, usually no normalization is needed, as the non-gaussianity is taken care of by
the outcome distribution, offset and link functions. Currently, identity, inverse and log-link are implemented.
If multiple decay parameters eta are supplied (etas is a vector), tests are performed for all weight matrices
and the resulting p-values combined using the Cauchy combination rule.

gpParams must be a list of length 2 with names 'X' and 'Y', consisting of matrices
with rownames "mean", "nugget", "range" and "sigma", and column names as in X and Y.
This argument allows to pass parameters of the Gaussian processes estimated with other software
to perform the score test.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\seealso{
\link{MoransISingle}, \link{ModTtestSingle}, \link{GAMsSingle}, \link{GPsSingle}
}
