% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarSingle.R
\name{sbivarSingle}
\alias{sbivarSingle}
\title{Test for bivariate spatial association in a single image}
\usage{
sbivarSingle(
  X,
  Y,
  Cx,
  Ey,
  method = c("Moran's I", "GAMs", "Modified t-test", "GPs"),
  normX = c("none", "rel", "log"),
  normY = c("none", "rel", "log"),
  pseudoCount = 1e-08,
  etas = c(5e-06, 4e-05, 2e-04),
  findMaxW = FALSE,
  returnSEsMoransI = TRUE,
  families = list(X = gaussian(), Y = gaussian()),
  n_points_grid = 600,
  verbose = TRUE,
  variogramModels = c("Exp", "Lin"),
  width = cutoff/15,
  cutoff = sqrt(2)/3,
  wo = c("Gauss", "nn"),
  numNNs = c(4, 8, 24),
  GPmethod = c("REML", "ML"),
  gpParams,
  Quants = c(0.005, 0.5),
  numLscAlts = 5,
  optControl = lmeControl(opt = "optim", maxIter = 500, msMaxIter = 500, niterEM = 1000,
    msMaxEval = 1000),
  corStruct = corGaus(form = ~x + y, nugget = TRUE, value = c(1, 0.25))
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{method}{A character string, indicating which method to apply}

\item{normX, normY, pseudoCount}{Normalization parameters, passed onto \link{normMat}}

\item{n_points_grid, families}{Passed onto \link{GAMsSingle}}

\item{verbose}{Should info on type of analysis be printed?}

\item{wo, variogramModels, numNNs, etas, cutoff, width, returnSEsMoransI, findMaxW}{Parameters for the calculation of Moran's I, passed onto \link{buildWeightMat}}

\item{GPmethod, Quants, numLscAlts, optControl, corStruct, gpParams}{Passed onto \link{fitGP}}
}
\value{
A list with at least the following components
\item{result}{A matrix which contains at least a p-values ("pVal") and a
Benjamini-Hochberg adjusted p-value ("pAdj"), sorted by increasing p-value.}
\item{families}{As provided}
\item{method}{As provided}
\item{multi}{FALSE, a flag for the type of analysis}
}
\description{
The tests can be applied to either disjoint or joint coordinate sets.
}
\details{
If only Cx is supplied and X and Y have the same number of rows, a joint analysis is performed
If Cx and Ey are provided, and X and Y have the same number of rows, equality of Cx and Ey is checked.
If true, a joint analysis is run, with a warning.

For GAMs, usually no normalization is needed, as the non-gaussianity is taken care of by
the outcome distribution, offset and link functions. Currently, identity, inverse and log-link are implemented.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\seealso{
\link{MoransISingle}, \link{ModTtestSingle}, \link{GAMsSingle}, \link{GPsSingle}
}
