% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarSingle.R
\name{sbivarSingle}
\alias{sbivarSingle}
\title{Test for bivariate spatial association in a single image}
\usage{
sbivarSingle(
  X,
  Y,
  Cx,
  Ey,
  method = c("GAMs", "Modified t-test", "GPs"),
  n_points_grid = 600,
  mapToFinest = FALSE,
  families = list(X = gaussian(), Y = gaussian()),
  GPmethod = c("REML", "ML", "gpytorch"),
  device = c("cpu", "cuda"),
  training_iter = 100L,
  gpParams,
  Quants = c(0.005, 0.5),
  numLscAlts = 10,
  optControl = lmeControl(opt = "optim", maxIter = 500, msMaxIter = 500, niterEM = 1000,
    msMaxEval = 1000),
  corStruct = corGaus(form = ~x + y, nugget = TRUE, value = c(1, 0.25))
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{method}{A character string, indicating which method to apply}

\item{n_points_grid, families}{Passed onto \link{wrapGAMs}}

\item{mapToFinest}{Passed onto \link{wrapModTtest}}

\item{GPmethod, device, training_iter, Quants, numLscAlts, optControl, corStruct}{Passed onto \link{fitGP}}

\item{gpParams}{Parameters of the Gaussian processes, see details}
}
\value{
A matrix which contains at least a p-values ("pVal") and a Benjamini-Hochberg adjusted p-value ("pAdj"),
sorted by increasing p-value.
}
\description{
The tests can be applied to either disjoint or joint coordinate sets.
}
\details{
If only Cx is supplied and X and Y have the same number of rows, a joint analysis is performed
If Cx and Ey are provided, and X and Y have the same number of rows, equality of Cx and Ey is checked.
If true, a joint analysis is run, with a warning.

Any normalization of the data should happen prior to calling this function.
For instance, count data or metabolome data are best scaled to relative values and log-normalized prior to fitting GPs.
For GAMs, usually no normalization is needed, as the non-gaussianity is taken care of by
the outcome distribution, offset and link functions. Currently, identity, inverse and log-link are implemented.

gpParams must be a list of length 2 with names 'X' and 'Y', consisting of matrices
with rownames "mean", "nugget", "range" and "sigma", and column names as in X and Y.
This argument allows to pass parameters of the Gaussian processes estimated with other software
to perform the score test.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
\examples{
n=7e1;m=9e1;p=3;k=4
X = matrix(rnorm(n*p), n, p, dimnames = list(NULL, paste0("X", seq_len(p))))
Y = matrix(rnorm(m*k), m, k, dimnames = list(NULL, paste0("Y", seq_len(k))))
Cx = matrix(runif(n*2), n, 2)
Ey = matrix(runif(m*2), m, 2)
colnames(Cx) = colnames(Ey) = c("x", "y")
resGAMs = sbivarSingle(X, Y, Cx, Ey, method = "GAMs")
resModtTest = sbivarSingle(X, Y, Cx, Ey, method = "Modified")
resModtTestJoint = sbivarSingle(X, Y[seq_len(nrow(X)),], Cx, method = "Modified")
resModtGPs = sbivarSingle(X, Y, Cx, Ey, method = "GPs")
}
