% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarSingle.R
\name{sbivarSingle}
\alias{sbivarSingle}
\title{Apply one or more tests for bivariate spatial association}
\usage{
sbivarSingle(
  X,
  Y,
  Cx,
  Ey,
  method = c("GAMs", "Modified t-test", "GPs"),
  n_points_grid = 500,
  mapToFinest = FALSE,
  families = list(X = gaussian(), Y = gaussian()),
  gpParams
)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{method}{A character string, indicating which method to apply}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}

\item{mapToFinest}{A boolean, should the one-to-one mapping for modified t-test
occur to the dataset with the best resolution?}

\item{families}{A vector of length 2 giving outcome values.}

\item{gpParams}{Parameters of the Gaussian processes}
}
\value{
A matrix which contains at least a p-values ("pVal") and a Benjamini-Hochberg adjusted p-value ("pAdj"),
sorted by increasing p-value.
}
\description{
The tests can be applied to either disjoint or joint coordinate sets.
}
\details{
If only Cx is supplied and X and Y have the same number of rows, a joint analysis is performed
If Cx and Ey are provided, and X and Y have the same number of rows, equality of Cx and Ey is checked.
If true, a joint analysis is run, with a warning.
}
\examples{
n=1e2;m=2e2;p=10;k=5
X = matrix(rnorm(n*p), n, p, dimnames = list(NULL, paste0("X", seq_len(p))))
Y = matrix(rnorm(m*k), m, k, dimnames = list(NULL, paste0("Y", seq_len(k))))
Cx = matrix(runif(n*2), n, 2)
Ey = matrix(runif(m*2), m, 2)
colnames(Cx) = colnames(Ey) = c("x", "y")
resGAMs = sbivarSingle(X, Y, Cx, Ey, method = "GAMs")
resModtTest = sbivarSingle(X, Y, Cx, Ey, method = "Modified")
resModtTestJoint = sbivarSingle(X, Y[seq_len(nrow(X)),], Cx, method = "Modified")
}
