% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbivarSinglePPP.R
\name{sbivarSinglePPP}
\alias{sbivarSinglePPP}
\title{Test for bivariate spatial association between a point pattern and quantitative spatial omics modality in a single image}
\usage{
sbivarSinglePPP(
  X,
  Y,
  Ey,
  method = c("Moran's I", "GAMs"),
  normY = c("none", "rel", "log"),
  pseudoCount = 1e-08,
  etas = c(5e-06, 5e-05, 2e-04),
  returnSEsMoransI = TRUE,
  families = list(Y = gaussian()),
  Gamm = FALSE,
  featuresX = getFeaturesX(X),
  featuresY = colnames(Y),
  n_points_grid = 600,
  verbose = TRUE,
  variogramModels = c("Exp", "Lin"),
  width = cutoff/15,
  cutoff = sqrt(2)/3,
  wo = c("Gauss", "nn"),
  numNNs = c(4, 8, 24),
  correlation = corGaus(form = ~x + y, nugget = TRUE, value = c(0.9 * max(apply(Ey, 2,
    function(x) diff(range(x)))), 0.25))
)
}
\arguments{
\item{X}{A marked point pattern of class \link[spatstat.geom]{ppp}}

\item{Y}{Matrix of omics measurements for the second, quantitative modality}

\item{Ey}{Coordinate matrix of dimension two, belonging to Y}

\item{method}{A character string, indicating which method to apply}

\item{normY, pseudoCount}{Normalization parameters, passed onto \link{normMat}}

\item{featuresX, featuresY}{Features to be tested. Defaults to all features, but specifying them allows to test a limited feature set,
while using the whole matrix to calculate library sizes as offset or for normalization for Y.}

\item{n_points_grid, families, Gamm}{Passed onto \link{GAMsSingle} for the second modality fitting}

\item{verbose}{Should info on type of analysis be printed?}

\item{wo, variogramModels, numNNs, etas, cutoff, width, returnSEsMoransI}{Parameters for the calculation of Moran's I, passed onto \link{buildWeightMat}}

\item{correlation}{Correlation structure, passed onto \link{fitGAM}}
}
\value{
A list with at least the following components
\item{result}{A matrix which contains at least a p-values ("pVal") and a
Benjamini-Hochberg adjusted p-value ("pAdj"), sorted by increasing p-value.}
\item{multi}{FALSE, a flag for the type of analysis}
\item{method,normX,normY}{As provided}
\item{families,wo,wParams}{Optional, as provided. wParams are either etas or numNNs}
}
\description{
This test is meant for the combination of a point pattern with a quantitative outcome with fixed grid location.
For two quantitative omics types, see \link{sbivarSingle}. For testing between point patterns, see e.g. \link[smoppix]{smoppix}
}
\details{
Ey must have rownames matching those in Y, and have two columns.
For GAMs, usually no normalization is needed, as the non-gaussianity is taken care of by
the outcome distribution, offset and link functions. Currently, identity, inverse and log-link are implemented.
}
\note{
All methods use multithreading on the cluster provided using the BiocParallel package
}
