% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGAM.R
\name{testGAM}
\alias{testGAM}
\title{Test for correlation between the predictions of two GAM models}
\usage{
testGAM(modelx, modely, predx, predy)
}
\arguments{
\item{modelx, modely}{Two fitted GAMs}

\item{predx, predy}{Predictions and covariance matrices of fitted GAMs in common grid}
}
\value{
A vector with correlation, its standard error and the p-value
}
\description{
Test for correlation between the predictions of two GAM models
}
