% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGP.R
\name{testGP}
\alias{testGP}
\title{Perform a score test on the bivariate spatial association in a Gaussian process.}
\usage{
testGP(x, y, Cx, Ey, altSigmas, distMat, solXonly, solYonly, sx)
}
\arguments{
\item{x, y}{outcome vectors}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{altSigmas}{A prepared series of bivariate association matrices}

\item{distMat}{The complete distance matrix of Cx and Ey stacked}

\item{solXonly, solYonly}{Parameters of the Gaussian processes of x and y}

\item{sx}{Inverse of the covariance matrix of x. Will be calculated if missing.}
}
\value{
A vector of length 2: a p-value and an indicator of the sign:
+1 for positive association, -1 for negative
}
\description{
This function tests for the variance of a random effect, causing the covariance, to be zero.
It is a score test as developed by \insertCite{Zhang2003}{sbivar}, with the test statistic
having a scaled chi-square distribution.
}
\details{
To tests are performed, one for positive and one for negative association,
and two times the smallest p-value to achieve a two-sided test. The sign indicates which direction was most significant.
}
\references{
\insertAllCited{}
}
