% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGP.R
\name{testGP}
\alias{testGP}
\title{Perform the score test for significance of the bivariate spatial association for Gaussian processes.}
\usage{
testGP(x, y, Cx, Ey, altSigmas, distMat, solXonly, solYonly, sx)
}
\arguments{
\item{x, y}{outcome vectors}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{altSigmas}{A prepared series of bivariate association matrices}

\item{distMat}{The distance matrix of Cx and Ey}

\item{solXonly, solYonly}{Parametersof the Gaussian process}

\item{sx}{Inverse of the covariance matrix of x. Will be calculated if missing.}
}
\value{
A vector of length 2: a p-value and an indicator of the sign:
+1 for positive association, -1 for negative
}
\description{
Perform the score test for significance of the bivariate spatial association for Gaussian processes.
}
