% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{topPairs}
\alias{topPairs}
\title{Return the top results found by sbivar}
\usage{
topPairs(x, ...)
}
\arguments{
\item{sbivarResults}{The sbivarResults object}
}
\value{
A dataframe of top results
}
\description{
Return the top results found by sbivar
}
