% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods-topPairs.R
\name{topPairs}
\alias{topPairs}
\alias{topPairs,sbivarResults-method}
\title{Return the top results found by sbivar}
\usage{
topPairs(x, ...)

\S4method{topPairs}{sbivarResults}(x, parameter = "Intercept", ...)
}
\arguments{
\item{x}{The sbivarResults object}

\item{...}{passed onto \link[utils]{head}}

\item{parameter}{For mulit-image analysis, the parameter for which top results should be shown}
}
\value{
A dataframe of top results
}
\description{
Return the top results found by sbivar
}
\examples{
example(sbivar, "sbivar")
topPairs(resGAMs)
}
