% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovPredGam.R
\name{vcovPredGam}
\alias{vcovPredGam}
\title{Return predictions of a gam, along with their variance-covariance matrix}
\usage{
vcovPredGam(model, newdata)
}
\arguments{
\item{model}{The gam model}

\item{newdata}{The grid on which new predictions are made}
}
\value{
A list with components
\item{pred}{A vector of predictions}
\item{vcov}{Variance-covariance matrix of the predictions}
}
\description{
The spline predictions are found as X *coef with X the spline basis and coef the spline coefficients.
The covariance matrix of the predictions is found using the formula:
Cov(pred) = X * Var(coef) * X',
where X is the basis matrix (design matrix for smooth terms), and Var(coef)
is the covariance matrix of the coefficients.
}
\seealso{
\link[mgcv]{vcov.gam}, \link[mgcv]{predict.gam}
}
