% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovPredGam.R
\name{vcovPredGam}
\alias{vcovPredGam}
\title{Return predictions of a gam, along with their variance-covariance matrix}
\usage{
vcovPredGam(model, newdata)
}
\arguments{
\item{model}{The gam model}

\item{newdata}{The grid on which new predictions are done}
}
\value{
A list with components
\item{pred}{A vector of predictions}
\item{vcov}{Variance-covariance matrix of the predictions}
}
\description{
The covariance matrix of the predictions is found using the formula:
Cov(pred) = X * Var(coef) * X'
Where X is the basis matrix (design matrix for smooth terms), and Var(coef) is the covariance matrix of the coefficients
From help(predict.gam): When type="lpmatrix" then a matrix is returned which
yields the values of the linear predictor (minus any offset) when postmultiplied by
the parameter vector (in this case se.fit is ignored). The latter option is most useful
for getting variance estimates for quantities derived from the model: for example integrated quantities, or derivatives of smooths.
}
\seealso{
\link[mgcv]{vcov.gam}, \link[mgcv]{predict.gam}
}
