% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapCorrelationsMulti.R
\name{wrapCorrelationsMulti}
\alias{wrapCorrelationsMulti}
\title{Find all cross-correlations for a list of matrices}
\usage{
wrapCorrelationsMulti(Xl, Yl, Cxl, Eyl, mapToFinest, jointCoordinates)
}
\arguments{
\item{Xl, Yl}{Lists of matrices of omics measurements}

\item{Cxl, Eyl}{Lists of corresponding coordinate matrices of dimension two}

\item{mapToFinest}{A boolean, should the one-to-one mapping for modified t-test
occur to the dataset with the best resolution?}

\item{jointCoordinates}{A boolean, are measurements on the same location
.}
}
\value{
A list of named correlation vectors
}
\description{
Find all raw cross-correlations between lists of observations matrices from different modalities.
Observations are matched to nearest neighbours if necessary
}
\seealso{
\link{matchCoords}
}
