% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGAM.R
\name{wrapGAMs}
\alias{wrapGAMs}
\title{Wrapper function to fit GAMs and test for all possible combinations}
\usage{
wrapGAMs(X, Y, Cx, Ey, families, n_points_grid)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{families}{A vector of length 2 giving outcome values.}

\item{n_points_grid}{The number of points in the new grid for the GAMs to be
evaluated on.}
}
\value{
see \link{testManyGAMs}
}
\description{
Wrapper function to fit GAMs and test for all possible combinations
}
