% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapGPs.R
\name{wrapGPs}
\alias{wrapGPs}
\title{Fit Gaussian processes (GPs) if needed, and perform score tests}
\usage{
wrapGPs(X, Y, Cx, Ey, gpParams)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{gpParams}{Parameters of the Gaussian processes}
}
\description{
Fit Gaussian processes (GPs) if needed, and perform score tests
}
