% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapModTtest.R
\name{wrapModTtest}
\alias{wrapModTtest}
\title{Perform nearest-neighbour matching, and apply all pairwise tests}
\usage{
wrapModTtest(X, Y, Cx, Ey, mapToFinest = FALSE)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{mapToFinest}{A boolean, should the one-to-one mapping for modified t-test
occur to the dataset with the best resolution?}
}
\value{
A dataframe of results sorted by p-value
}
\description{
Perform nearest-neighbour matching, and apply all pairwise tests
}
