% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapModTtest.R
\name{wrapModTtest}
\alias{wrapModTtest}
\title{Perform nearest-neighbour matching if necessary, and apply all pairwise tests.}
\usage{
wrapModTtest(X, Y, Cx, Ey, mapToFinest = FALSE, jointCoordinates = FALSE)
}
\arguments{
\item{X, Y}{Matrices of omics measurements}

\item{Cx, Ey}{Corresponding coordinate matrices of dimension two}

\item{mapToFinest}{A boolean, should the one-to-one mapping for modified t-test
occur to the dataset with the best resolution?}

\item{jointCoordinates}{A boolean, are measurements on the same location
.}
}
\value{
A dataframe of results sorted by p-value, also containing effective sample size (ESS) and correlation estimate.
}
\description{
If measurements are not on the same location, they are matched using nearest neighbour matching with the \link[RANN]{nn2} function.
Then modified t-test is applied to all pairs, which tests for the significance
of the Pearson correlation while accounting for spatial autocorrelation \insertCite{Clifford1989}{sbivar}.
}
\seealso{
\link[SpatialPack]{modified.ttest}
}
